+++ src/Features/VR.cpp
@@ -1,39 +1,20 @@
 ﻿#include "VR.h"
 #include "FeatureConstraints.h"
 #include "Menu.h"
-#include "Menu/Fonts.h"
 #include "RE/B/BSOpenVR.h"
-#include "RE/N/NiPoint3.h"
 #include "RE/P/PlayerCharacter.h"
 #include "Upscaling.h"
-#include <openvr.h>
+#include "VR/OpenVRDetection.h"
 
 #include "State.h"
 #include "Utils/D3D.h"
-#include "Utils/PerfUtils.h"
-#include "Utils/UI.h"
 #include "Utils/VRUtils.h"
-#include <DirectXMath.h>
-#include <SimpleMath.h>
-#include <chrono>
-#include <cmath>
+
 #include <d3d11.h>
-#include <d3dcompiler.h>
 #include <imgui_impl_dx11.h>
-#include <unordered_map>
-#include <windows.h>
-#include <winver.h>
-#pragma comment(lib, "version.lib")
+#include <openvr.h>
 
 using AttachMode = VR::Settings::OverlayAttachMode;
-
-namespace
-{
-	bool BeginTabItemWithFont(const char* label, Menu::FontRole role, ImGuiTabItemFlags flags = ImGuiTabItemFlags_None)
-	{
-		return MenuFonts::BeginTabItemWithFont(label, role, flags);
-	}
-}
 
 NLOHMANN_DEFINE_TYPE_NON_INTRUSIVE_WITH_DEFAULT(
 	VR::Settings,
@@ -70,7 +51,6 @@ NLOHMANN_DEFINE_TYPE_NON_INTRUSIVE_WITH_DEFAULT(
 void VR::LoadSettings(json& o_json)
 {
 	settings = o_json.get<Settings>();
-	// Validate and clamp loaded settings to ensure they're within valid ranges
 	settings.ClampToValidRanges();
 }
 
@@ -86,10 +66,8 @@ void VR::RestoreDefaultSettings()
 
 void VR::SetupResources()
 {
-	// Detect OpenVR version and compatibility early to avoid CTDs
 	DetectOpenVRInfo();
 
-	// Log OpenVR information
 	if (openVRInfo.isAvailable) {
 		logger::info("OpenVR DLL detected:");
 		logger::info("  Path: {}", openVRInfo.dllPath);
@@ -119,7 +97,7 @@ void VR::PostPostLoad()
 {
 	gDepthBufferCulling = reinterpret_cast<bool*>(REL::Offset(0x1EC6B88).address());
 	if (!gDepthBufferCulling) {
-		static bool s_defaultDepthBufferCulling = false;  // safe fallback
+		static bool s_defaultDepthBufferCulling = false;
 		gDepthBufferCulling = &s_defaultDepthBufferCulling;
 		logger::warn("VR: gDepthBufferCulling address not found - using fallback default (false)");
 	}
@@ -131,8 +109,7 @@ void VR::PostPostLoad()
 		logger::warn("VR: gMinOccludeeBoxExtent address not found - using fallback default (10.0)");
 	}
 
-	// Migration: Fix legacy overlay keybinds that were deserialized incorrectly
-	// Plain integer values (e.g., 32) deserialize as Keyboard device instead of VR controller
+	// Migration: Fix legacy overlay keybinds
 	if (settings.VROverlayCloseKeys.size() == 1) {
 		auto& closeKey = settings.VROverlayCloseKeys[0];
 		if (closeKey.GetDevice() == ControllerDevice::Keyboard && closeKey.GetKey() == 32) {
@@ -148,7 +125,6 @@ void VR::PostPostLoad()
 		}
 	}
 
-	// Patches BSGeometry::CopyTransformAndBounds to copy the model-bound translation across correctly instead of overwriting it with the bounding sphere centre
 	REL::safe_write(REL::RelocationID(0, 0, 69528).address() + REL::Relocate(0, 0, 0xD9) + 0x2, 0x148);
 	REL::safe_write(REL::RelocationID(0, 0, 69528).address() + REL::Relocate(0, 0, 0xE5) + 0x2, 0x14C);
 	REL::safe_write(REL::RelocationID(0, 0, 69528).address() + REL::Relocate(0, 0, 0xF1) + 0x2, 0x150);
@@ -156,8 +132,6 @@ void VR::PostPostLoad()
 
 void VR::DataLoaded()
 {
-	// Initialize occlusion culling based on settings, but force-disable if an external
-	// upscaler is active (FSR/DLSS) since upscalers may modify the depth buffer.
 	bool desired = settings.EnableDepthBufferCullingExterior;
 	UpdateDepthBufferCulling(desired, { "VR", "EnableDepthBufferCullingExterior" });
 
@@ -170,8 +144,6 @@ void VR::DataLoaded()
 
 void VR::EarlyPrepass()
 {
-	// Respect user settings unless an external upscaler is active; if so, force-disable
-	// depth-buffer culling to avoid incorrect occlusion tests in VR.
 	bool isInterior = RE::TES::GetSingleton()->interiorCell != nullptr;
 	auto settingId = isInterior ? FeatureConstraints::SettingId{ "VR", "EnableDepthBufferCullingInterior" } : FeatureConstraints::SettingId{ "VR", "EnableDepthBufferCullingExterior" };
 	bool desired = isInterior ? settings.EnableDepthBufferCullingInterior : settings.EnableDepthBufferCullingExterior;
@@ -179,1145 +151,16 @@ void VR::EarlyPrepass()
 }
 
 //=============================================================================
-// OVERLAY FEATURE OVERRIDES
-//=============================================================================
-
-void VR::DrawOverlay()
-{
-	auto& vr = globals::features::vr;
-	if (!vr.IsOpenVRCompatible())
-		return;
-	static LARGE_INTEGER overlayShowStart = { 0 };
-	static LARGE_INTEGER freq = { 0 };
-
-	bool shouldShow = settings.kAutoHideSeconds > 0 && globals::game::ui && globals::game::ui->IsMenuOpen(RE::MainMenu::MENU_NAME) && globals::menu && !globals::menu->IsEnabled;
-
-	if (!shouldShow) {
-		overlayShowStart.QuadPart = 0;  // Reset timer when overlay is not shown
-		return;
-	}
-
-	if (freq.QuadPart == 0) {
-		QueryPerformanceFrequency(&freq);
-	}
-
-	LARGE_INTEGER now;
-	QueryPerformanceCounter(&now);
-
-	if (overlayShowStart.QuadPart == 0) {
-		overlayShowStart = now;
-	}
-
-	double elapsed = double(now.QuadPart - overlayShowStart.QuadPart) / double(freq.QuadPart);
-	const double autoHideSeconds = static_cast<double>(settings.kAutoHideSeconds);
-	if (elapsed >= autoHideSeconds) {
-		return;
-	}
-	int secondsLeft = int(std::ceil(autoHideSeconds - elapsed));
-
-	ImGuiIO& io = ImGui::GetIO();
-	ImVec2 overlaySize(520, 0);  // width, height auto
-	// Center horizontally and vertically for better VR readability
-	ImVec2 overlayPos = ImVec2((io.DisplaySize.x - overlaySize.x) * 0.5f, (io.DisplaySize.y * 0.35f));
-	ImGui::SetNextWindowPos(overlayPos, ImGuiCond_Always);
-	ImGui::SetNextWindowSize(overlaySize, ImGuiCond_Always);
-	ImGui::SetNextWindowBgAlpha(0.95f);
-
-	ImGui::Begin("HowToUseOverlay", nullptr, ImGuiWindowFlags_NoTitleBar | ImGuiWindowFlags_AlwaysAutoResize | ImGuiWindowFlags_NoSavedSettings | ImGuiWindowFlags_NoFocusOnAppearing | ImGuiWindowFlags_NoNav);
-
-	ImGui::PushTextWrapPos(ImGui::GetCursorPos().x + 500.0f);
-	ImGui::TextWrapped("How to Use VR Community Shaders Menu:");
-	ImGui::Separator();
-	ImGui::TextWrapped("You must open the Main Menu or Tween Menu before VR controls work.");
-	ImGui::Spacing();
-	ImGui::PopTextWrapPos();
-
-	ImGui::Text("Open Menu: ");
-	ImGui::SameLine();
-	Util::DrawButtonCombo(settings.VRMenuOpenKeys, true);
-
-	ImGui::Text("Close Menu: ");
-	ImGui::SameLine();
-	Util::DrawButtonCombo(settings.VRMenuCloseKeys, true);
-
-	ImGui::Spacing();
-	ImGui::PushTextWrapPos(ImGui::GetCursorPos().x + 500.0f);
-	ImGui::TextWrapped("Grip + Thumbstick: Adjust overlay depth (closer/farther)");
-	ImGui::Spacing();
-	ImGui::TextWrapped("Tip: Disable this VR overlay by setting Attach Mode to 'None' in VR settings.");
-	ImGui::Spacing();
-	ImGui::TextWrapped("(This welcome message will auto-hide in %d seconds)", secondsLeft);
-	ImGui::TextWrapped("(Disable in: VR settings > Controller Input Instructions)");
-	ImGui::PopTextWrapPos();
-
-	ImGui::End();
-}
-
-namespace
-{
-	void DrawControllerInputInstructions();
-	void DrawGeneralVRSettings();
-	void DrawMenuSettings();
-	void DrawMouseSettings();
-	void DrawDragSettings();
-	void DrawKeyBindings();
-	void DrawDebugSection();
-}
-
-void VR::DrawSettings()
-{
-	auto menu = globals::menu;
-	if (!menu)
-		return;
-	if (ImGui::BeginTabBar("##VRTabs", ImGuiTabBarFlags_None)) {
-		// General Settings Tab
-		if (BeginTabItemWithFont("General", Menu::FontRole::Subheading)) {
-			if (ImGui::BeginChild("##VRGeneralFrame", { 0, 0 }, true)) {
-				DrawGeneralVRSettings();
-				DrawControllerInputInstructions();
-				DrawMenuSettings();
-				DrawMouseSettings();
-				DrawDragSettings();
-			}
-			ImGui::EndChild();
-			ImGui::EndTabItem();
-		}
-
-		// Key Bindings Tab
-		if (IsOpenVRCompatible()) {
-			if (BeginTabItemWithFont("Bindings", Menu::FontRole::Subheading)) {
-				if (ImGui::BeginChild("##VRBindingsFrame", { 0, 0 }, true)) {
-					DrawKeyBindings();
-				}
-				ImGui::EndChild();
-				ImGui::EndTabItem();
-			}
-		}
-		// Debug Tab (existing debug functionality)
-		if (BeginTabItemWithFont("Debug", Menu::FontRole::Subheading)) {
-			if (ImGui::BeginChild("##VRDebugFrame", { 0, 0 }, true)) {
-				DrawDebugSection();
-			}
-			ImGui::EndChild();
-			ImGui::EndTabItem();
-		}
-
-		ImGui::EndTabBar();
-	}
-
-	// Combo recording popup
-	if (this->isCapturingCombo) {
-		ImGui::OpenPopup("Record Combo");
-		ImGui::SetNextWindowSize(ImVec2(400, 200), ImGuiCond_FirstUseEver);
-		if (ImGui::BeginPopupModal("Record Combo", nullptr, ImGuiWindowFlags_AlwaysAutoResize)) {
-			// Helper function to get button name
-			auto GetButtonName = [](uint32_t key) -> const char* {
-				switch (key) {
-				case static_cast<uint32_t>(RE::BSOpenVRControllerDevice::Keys::kTrigger):
-					return "Trigger";
-				case static_cast<uint32_t>(RE::BSOpenVRControllerDevice::Keys::kGrip):
-					return "Grip";
-				case static_cast<uint32_t>(RE::BSOpenVRControllerDevice::Keys::kTouchpadClick):
-					return "Touchpad";
-				case static_cast<uint32_t>(RE::BSOpenVRControllerDevice::Keys::kJoystickTrigger):
-					return "Stick Click";
-				case static_cast<uint32_t>(RE::BSOpenVRControllerDevice::Keys::kXA):
-					return "A/X";
-				case static_cast<uint32_t>(RE::BSOpenVRControllerDevice::Keys::kBY):
-					return "B/Y";
-				default:
-					return "Unknown";
-				}
-			};
-
-			ImGui::Text("Recording combo for: %s", this->currentComboName ? this->currentComboName : "Unknown");
-			ImGui::Spacing();
-
-			ImGui::TextDisabled("(During recording, any controller's buttons can be used. Requirement is only enforced during use.)");
-
-			ImGui::Spacing();
-
-			// Show countdown timer with color
-			double remainingTime = this->comboTimeout - (Util::GetNowSecs() - this->comboStartTime);
-			ImVec4 timerColor = remainingTime > 2.0 ? Util::Colors::GetTimerGood() :
-			                    remainingTime > 1.0 ? Util::Colors::GetTimerWarning() :
-			                                          Util::Colors::GetTimerCritical();
-			ImGui::TextColored(timerColor, "Time remaining: %.1f seconds", remainingTime);
-
-			ImGui::Spacing();
-
-			// Show recorded buttons
-			if (this->recordedCombo.empty()) {
-				ImGui::Text("Press buttons to record combo...");
-			} else {
-				ImGui::Text("Recorded buttons:");
-				// Create a sorted list of decoded buttons for consistent display
-				std::vector<ButtonCombo> sortedRecordedCombos;
-				for (size_t i = 0; i < this->recordedCombo.size(); ++i) {
-					sortedRecordedCombos.push_back(this->recordedCombo[i]);
-				}
-				std::sort(sortedRecordedCombos.begin(), sortedRecordedCombos.end(),
-					[](const ButtonCombo& a, const ButtonCombo& b) {
-						return a.GetKey() < b.GetKey();
-					});
-
-				Util::DrawButtonCombo(sortedRecordedCombos, false);
-			}
-
-			ImGui::Spacing();
-			ImGui::Separator();
-			ImGui::Spacing();
-
-			// Instructions
-			ImGui::Text("Press ENTER to accept, ESC to cancel");
-
-			// Handle button recording
-			// Check for VR controller button presses - record them (any controller allowed during recording)
-			bool buttonPressed = false;
-			uint32_t pressedKey = 0;
-			ControllerDevice pressedDevice = ControllerDevice::Both;  // Default to Both, will set below
-
-			// Check primary controller buttons
-			for (const auto& [keyCode, buttonState] : primaryControllerState.GetActiveButtons()) {
-				if (buttonState->isPressed) {
-					pressedKey = keyCode;
-					buttonPressed = true;
-					pressedDevice = ControllerDevice::Primary;
-					break;
-				}
-			}
-
-			// Check secondary controller buttons if primary didn't have any
-			if (!buttonPressed) {
-				for (const auto& [keyCode, buttonState] : secondaryControllerState.GetActiveButtons()) {
-					if (buttonState->isPressed) {
-						pressedKey = keyCode;
-						buttonPressed = true;
-						pressedDevice = ControllerDevice::Secondary;
-						break;
-					}
-				}
-			}
-
-			// Record button press
-			if (buttonPressed) {
-				// Check if this button is already in the combo (avoid duplicates)
-				auto it = recordingButtonControllers.find(pressedKey);
-				if (it == recordingButtonControllers.end()) {
-					// Not yet recorded, add with the current device
-					recordingButtonControllers[pressedKey] = pressedDevice;
-				} else {
-					// Already recorded, if the other controller is now pressed, set to BOTH
-					if (it->second != pressedDevice && it->second != ControllerDevice::Both) {
-						it->second = ControllerDevice::Both;
-					}
-				}
-				// Update the recordedCombo vector to match the map
-				this->recordedCombo.clear();
-				for (const auto& [key, device] : recordingButtonControllers) {
-					this->recordedCombo.push_back(ButtonCombo(device, key));
-				}
-			}
-
-			// Handle ENTER key to accept combo
-			if (ImGui::IsKeyPressed(ImGui::GetKeyIndex(ImGuiKey_Enter)) || ImGui::IsKeyPressed(ImGui::GetKeyIndex(ImGuiKey_KeypadEnter))) {
-				if (!this->recordedCombo.empty()) {
-					// Apply the recorded combo to the correct settings vector
-					switch (this->currentComboType) {
-					case VR::ComboType::MenuOpen:
-						settings.VRMenuOpenKeys = this->recordedCombo;
-						break;
-					case VR::ComboType::MenuClose:
-						settings.VRMenuCloseKeys = this->recordedCombo;
-						break;
-					case VR::ComboType::OverlayOpen:
-						settings.VROverlayOpenKeys = this->recordedCombo;
-						break;
-					case VR::ComboType::OverlayClose:
-						settings.VROverlayCloseKeys = this->recordedCombo;
-						break;
-					default:
-						break;
-					}
-				}
-
-				// Reset recording state
-				this->isCapturingCombo = false;
-				this->currentComboType = VR::ComboType::None;
-				this->currentComboName = nullptr;
-				this->recordedCombo.clear();
-				this->comboStartTime = 0.0;
-				recordingButtonControllers.clear();
-				ImGui::CloseCurrentPopup();
-			}
-
-			// Handle ESC key to cancel
-			if (ImGui::IsKeyPressed(ImGui::GetKeyIndex(ImGuiKey_Escape))) {
-				// Reset recording state
-				this->isCapturingCombo = false;
-				this->currentComboType = VR::ComboType::None;
-				this->currentComboName = nullptr;
-				this->recordedCombo.clear();
-				this->comboStartTime = 0.0;
-				recordingButtonControllers.clear();
-				ImGui::CloseCurrentPopup();
-			}
-
-			// Handle timeout - auto-accept if buttons were pressed, auto-cancel if not
-			if (remainingTime <= 0.0) {
-				if (!this->recordedCombo.empty()) {
-					// Auto-accept if buttons were pressed - apply to correct settings vector
-					switch (this->currentComboType) {
-					case VR::ComboType::MenuOpen:
-						settings.VRMenuOpenKeys = this->recordedCombo;
-						break;
-					case VR::ComboType::MenuClose:
-						settings.VRMenuCloseKeys = this->recordedCombo;
-						break;
-					case VR::ComboType::OverlayOpen:
-						settings.VROverlayOpenKeys = this->recordedCombo;
-						break;
-					case VR::ComboType::OverlayClose:
-						settings.VROverlayCloseKeys = this->recordedCombo;
-						break;
-					default:
-						break;
-					}
-				}
-				// Auto-cancel if no buttons were pressed (do nothing, just close)
-
-				// Reset recording state
-				this->isCapturingCombo = false;
-				this->currentComboType = VR::ComboType::None;
-				this->currentComboName = nullptr;
-				this->recordedCombo.clear();
-				this->comboStartTime = 0.0;
-				recordingButtonControllers.clear();
-				ImGui::CloseCurrentPopup();
-			}
-
-			ImGui::EndPopup();
-		}
-	}
-}
-
-namespace
-{
-	void DrawControllerInputInstructions()
-	{
-		auto& vr = globals::features::vr;
-		auto& settings = vr.settings;
-		if (!vr.IsOpenVRCompatible())
-			return;
-		if (ImGui::CollapsingHeader("Controller Input Instructions", ImGuiTreeNodeFlags_DefaultOpen)) {
-			ImGui::SliderInt("Auto-hide Welcome overlay timeout", &settings.kAutoHideSeconds, 0, VR::Config::kMaxAutoHideSeconds,
-				settings.kAutoHideSeconds <= 0 ? "Hidden" : "%d seconds");
-			if (auto _tt = Util::HoverTooltipWrapper()) {
-				ImGui::Text("Set to 0 to hide the overlay, or a positive value to show it for that many seconds");
-			}
-			ImGui::TextWrapped("Menu (while in the main menu or tween menu):");
-			if (ImGui::BeginTable("MenuInstructionsTable", 2, ImGuiTableFlags_Borders | ImGuiTableFlags_RowBg)) {
-				ImGui::TableNextRow();
-				ImGui::TableSetColumnIndex(0);
-				ImGui::Text("Open Community Shaders Menu:");
-				ImGui::TableSetColumnIndex(1);
-				Util::DrawButtonCombo(settings.VRMenuOpenKeys, true);
-				ImGui::TableNextRow();
-				ImGui::TableSetColumnIndex(0);
-				ImGui::Text("Close Community Shaders Menu:");
-				ImGui::TableSetColumnIndex(1);
-				Util::DrawButtonCombo(settings.VRMenuCloseKeys, true);
-				ImGui::EndTable();
-			}
-			ImGui::TextWrapped("Overlay (while in the main menu or tween menu):");
-			if (ImGui::BeginTable("OverlayInstructionsTable", 2, ImGuiTableFlags_Borders | ImGuiTableFlags_RowBg)) {
-				ImGui::TableNextRow();
-				ImGui::TableSetColumnIndex(0);
-				ImGui::Text("Open Overlay:");
-				ImGui::TableSetColumnIndex(1);
-				Util::DrawButtonCombo(settings.VROverlayOpenKeys, true);
-				ImGui::TableNextRow();
-				ImGui::TableSetColumnIndex(0);
-				ImGui::Text("Close Overlay:");
-				ImGui::TableSetColumnIndex(1);
-				Util::DrawButtonCombo(settings.VROverlayCloseKeys, true);
-				ImGui::EndTable();
-			}
-			ImGui::TextWrapped("Menu Controller Input:");
-			if (ImGui::BeginTable("ControllerInputTable", 2, ImGuiTableFlags_Borders | ImGuiTableFlags_RowBg)) {
-				ImGui::TableNextRow();
-				ImGui::TableSetColumnIndex(0);
-				ImGui::TextColored(Util::GetControllerBothColor(), "Trigger (Both Controllers)");
-				ImGui::TableSetColumnIndex(1);
-				ImGui::Text("Left mouse button");
-				ImGui::TableNextRow();
-				ImGui::TableSetColumnIndex(0);
-				ImGui::TextColored(Util::GetControllerBothColor(), "Grip (Both Controllers)");
-				ImGui::TableSetColumnIndex(1);
-				ImGui::Text("Right mouse button");
-				ImGui::TableNextRow();
-				ImGui::TableSetColumnIndex(0);
-				ImGui::TextColored(Util::GetControllerBothColor(), "Touchpad Click (Both Controllers)");
-				ImGui::TableSetColumnIndex(1);
-				ImGui::Text("Middle mouse button");
-				ImGui::TableNextRow();
-				ImGui::TableSetColumnIndex(0);
-				ImGui::TextColored(Util::GetControllerBothColor(), "Stick Click (Both Controllers)");
-				ImGui::TableSetColumnIndex(1);
-				ImGui::Text("Middle mouse button");
-				ImGui::TableNextRow();
-				ImGui::TableSetColumnIndex(0);
-				ImGui::TextColored(Util::GetControllerBothColor(), "A/X (Both Controllers)");
-				ImGui::TableSetColumnIndex(1);
-				ImGui::Text("Enter");
-				ImGui::TableNextRow();
-				ImGui::TableSetColumnIndex(0);
-				ImGui::TextColored(Util::GetControllerPrimaryColor(), "B/Y (Primary Controller)");
-				ImGui::TableSetColumnIndex(1);
-				ImGui::Text("Tab");
-				ImGui::TableNextRow();
-				ImGui::TableSetColumnIndex(0);
-				ImGui::TextColored(Util::GetControllerSecondaryColor(), "B/Y (Secondary Controller)");
-				ImGui::TableSetColumnIndex(1);
-				ImGui::Text("Shift+Tab");
-				ImGui::EndTable();
-			}
-			// Thumbstick instructions
-			bool useAttachedControllerForCursor = (settings.attachMode == VR::Settings::OverlayAttachMode::ControllerOnly || settings.attachMode == VR::Settings::OverlayAttachMode::Both);
-			if (ImGui::BeginTable("ThumbstickInstructionsTable", 2, ImGuiTableFlags_Borders | ImGuiTableFlags_RowBg)) {
-				if (useAttachedControllerForCursor) {
-					if (settings.VRMenuAttachController == ControllerDevice::Primary) {
-						ImGui::TableNextRow();
-						ImGui::TableSetColumnIndex(0);
-						ImGui::TextColored(Util::GetControllerPrimaryColor(), "Primary Controller Thumbstick");
-						ImGui::TableSetColumnIndex(1);
-						ImGui::Text("Mouse movement (attached controller)");
-						ImGui::TableNextRow();
-						ImGui::TableSetColumnIndex(0);
-						ImGui::TextColored(Util::GetControllerSecondaryColor(), "Secondary Controller Thumbstick");
-						ImGui::TableSetColumnIndex(1);
-						ImGui::Text("Scroll");
-					} else {
-						ImGui::TableNextRow();
-						ImGui::TableSetColumnIndex(0);
-						ImGui::TextColored(Util::GetControllerPrimaryColor(), "Primary Controller Thumbstick");
-						ImGui::TableSetColumnIndex(1);
-						ImGui::Text("Scroll");
-						ImGui::TableNextRow();
-						ImGui::TableSetColumnIndex(0);
-						ImGui::TextColored(Util::GetControllerSecondaryColor(), "Secondary Controller Thumbstick");
-						ImGui::TableSetColumnIndex(1);
-						ImGui::Text("Mouse movement (attached controller)");
-					}
-				} else {
-					ImGui::TableNextRow();
-					ImGui::TableSetColumnIndex(0);
-					ImGui::TextColored(Util::GetControllerPrimaryColor(), "Primary Controller Thumbstick");
-					ImGui::TableSetColumnIndex(1);
-					ImGui::Text("Mouse movement (HMD mode)");
-					ImGui::TableNextRow();
-					ImGui::TableSetColumnIndex(0);
-					ImGui::TextColored(Util::GetControllerSecondaryColor(), "Secondary Controller Thumbstick");
-					ImGui::TableSetColumnIndex(1);
-					ImGui::Text("Scroll");
-				}
-				ImGui::EndTable();
-			}
-		}
-	}
-
-	void DrawGeneralVRSettings()
-	{
-		auto& vr = globals::features::vr;
-		VR::Settings& settings = vr.settings;
-		if (ImGui::CollapsingHeader("General Settings", ImGuiTreeNodeFlags_DefaultOpen)) {
-			// Use constraint-aware checkboxes that automatically handle disabling
-			// and showing tooltips when other features constrain these settings
-			Util::ConstrainedUI::Checkbox("Enable Depth Buffer Culling in Exteriors",
-				&settings.EnableDepthBufferCullingExterior,
-				{ "VR", "EnableDepthBufferCullingExterior" });
-			// Show normal tooltip when not constrained
-			auto exteriorConstraint = FeatureConstraints::GetConstraints({ "VR", "EnableDepthBufferCullingExterior" });
-			if (!exteriorConstraint.isConstrained) {
-				if (auto _tt = Util::HoverTooltipWrapper()) {
-					ImGui::Text("Improves performance in exteriors, recommended ON.");
-				}
-			}
-
-			Util::ConstrainedUI::Checkbox("Enable Depth Buffer Culling in Interiors",
-				&settings.EnableDepthBufferCullingInterior,
-				{ "VR", "EnableDepthBufferCullingInterior" });
-			// Show normal tooltip when not constrained
-			auto interiorConstraint = FeatureConstraints::GetConstraints({ "VR", "EnableDepthBufferCullingInterior" });
-			if (!interiorConstraint.isConstrained) {
-				if (auto _tt = Util::HoverTooltipWrapper()) {
-					ImGui::Text("Improves performance in interiors, recommended OFF due to occasional visual glitches.");
-				}
-			}
-
-			if (ImGui::SliderFloat("Min Occludee Box Extent", &settings.MinOccludeeBoxExtent, 0.0f, 1000.0f, "%.1f"))
-				*vr.gMinOccludeeBoxExtent = settings.MinOccludeeBoxExtent;
-			if (auto _tt = Util::HoverTooltipWrapper()) {
-				ImGui::Text("Minimum bounding box dimensions for object occlusion culling. Lower values improve performance but may result in visual artifacts.");
-			}
-		}
-	}
-
-	void DrawMenuSettings()
-	{
-		auto& vr = globals::features::vr;
-		auto& settings = vr.settings;
-		if (!vr.IsOpenVRCompatible())
-			return;
-		if (ImGui::CollapsingHeader("Menu Settings", ImGuiTreeNodeFlags_DefaultOpen)) {
-			float maxScale = VR::Config::kMaxMenuScale;
-			ImGui::SliderFloat("Menu Scale", &settings.VRMenuScale, VR::Config::kMinMenuScale, maxScale, "%.2f");
-			const char* positioningMethods[] = { "HMD Relative", "Fixed World Position" };
-			int prevMethod = settings.VRMenuPositioningMethod;
-			if (ImGui::Combo("Menu Positioning Method", &settings.VRMenuPositioningMethod, positioningMethods, IM_ARRAYSIZE(positioningMethods))) {
-				if (prevMethod != 1 && settings.VRMenuPositioningMethod == 1) {
-					vr.SetFixedOverlayToCurrentHMD();
-					auto player = RE::PlayerCharacter::GetSingleton();
-					if (player)
-						vr.savedPlayerWorldPos = player->GetPosition();
-				}
-			}
-			const char* attachModes[] = { "HMD Only", "Controller Only", "Both", "None (Disabled)" };
-			int attachModeInt = static_cast<int>(settings.attachMode);
-			if (ImGui::Combo("Attach Mode", &attachModeInt, attachModes, IM_ARRAYSIZE(attachModes))) {
-				settings.attachMode = static_cast<VR::Settings::OverlayAttachMode>(attachModeInt);
-			}
-
-			// Controller-specific settings (only show when controller mode is active)
-			if (settings.attachMode == VR::Settings::OverlayAttachMode::ControllerOnly ||
-				settings.attachMode == VR::Settings::OverlayAttachMode::Both) {
-				const char* attachControllers[] = { "Primary Controller", "Secondary Controller" };
-				int attachControllerInt = static_cast<int>(settings.VRMenuAttachController);
-				if (ImGui::Combo("Attach to Controller", &attachControllerInt, attachControllers, IM_ARRAYSIZE(attachControllers))) {
-					settings.VRMenuAttachController = static_cast<ControllerDevice>(attachControllerInt);
-				}
-
-				ImGui::Separator();
-				ImGui::Text("Controller Offset Settings");
-				ImGui::SliderFloat("Controller Offset X", &settings.VRMenuControllerOffsetX, -2.0f, 2.0f, "%.2f");
-				ImGui::SliderFloat("Controller Offset Y", &settings.VRMenuControllerOffsetY, -2.0f, 2.0f, "%.2f");
-				ImGui::SliderFloat("Controller Offset Z", &settings.VRMenuControllerOffsetZ, -2.0f, 2.0f, "%.2f");
-			}
-
-			// HMD-specific settings (only show when HMD mode is active)
-			if (settings.attachMode == VR::Settings::OverlayAttachMode::HMDOnly ||
-				settings.attachMode == VR::Settings::OverlayAttachMode::Both) {
-				ImGui::Separator();
-				ImGui::Text("HMD Offset Settings");
-				ImGui::SliderFloat("HMD Offset X", &settings.VRMenuOffsetX, -2.0f, 2.0f, "%.2f");
-				ImGui::SliderFloat("HMD Offset Y", &settings.VRMenuOffsetY, -2.0f, 2.0f, "%.2f");
-				ImGui::SliderFloat("HMD Offset Z", &settings.VRMenuOffsetZ, -4.0f, 1.0f, "%.2f");
-			}
-
-			// Fixed World Position: show auto reset distance and manual reset button
-			if (settings.VRMenuPositioningMethod == 1) {  // 1 = Fixed World Position
-				ImGui::Separator();
-				ImGui::Text("Fixed World Position Settings");
-				ImGui::SliderFloat("Auto Reset Distance (game units)", &settings.VRMenuAutoResetDistance, 100.0f, 5000.0f, "%.0f");
-				if (auto _tt = Util::HoverTooltipWrapper()) {
-					ImGui::Text("If you move farther than this distance from the menu, it will automatically reset to your HMD position. %s", Util::Units::FormatDistance(settings.VRMenuAutoResetDistance).c_str());
-				}
-				if (ImGui::Button("Reset Menu to HMD Position")) {
-					vr.SetFixedOverlayToCurrentHMD();
-				}
-			}
-		}
-	}
-
-	void DrawMouseSettings()
-	{
-		auto& vr = globals::features::vr;
-		if (!vr.IsOpenVRCompatible())
-			return;
-		VR::Settings& settings = vr.settings;
-		if (ImGui::CollapsingHeader("Input Settings", ImGuiTreeNodeFlags_DefaultOpen)) {
-			// Wand pointing settings
-			if (ImGui::Checkbox("Enable Wand Pointing", &settings.EnableWandPointing)) {
-				// Reset wand state when toggling
-				vr.wandState.isIntersecting = false;
-			}
-			if (auto _tt = Util::HoverTooltipWrapper()) {
-				ImGui::Text("Use controller ray-casting to point at UI elements");
-			}
-			ImGui::Separator();
-			ImGui::Text("Joystick Settings");
-			ImGui::SliderFloat("Mouse Deadzone", &settings.mouseDeadzone, 0.0f, 1.0f, "%.2f");
-			if (auto _tt = Util::HoverTooltipWrapper()) {
-				ImGui::Text("Thumbstick deadzone for joystick cursor movement");
-			}
-			ImGui::SliderFloat("Mouse Speed", &settings.mouseSpeed, 0.1f, 50.0f, "%.2f");
-			if (auto _tt = Util::HoverTooltipWrapper()) {
-				ImGui::Text("Speed multiplier for joystick cursor movement");
-			}
-		}
-	}
-
-	void DrawDragSettings()
-	{
-		auto& vr = globals::features::vr;
-		if (!vr.IsOpenVRCompatible())
-			return;
-		VR::Settings& settings = vr.settings;
-		if (ImGui::CollapsingHeader("Drag Settings", ImGuiTreeNodeFlags_DefaultOpen)) {
-			if (ImGui::CollapsingHeader("Drag Instructions", ImGuiTreeNodeFlags_DefaultOpen)) {
-				ImGui::TextWrapped("Overlay Positioning (Grip + Drag):");
-				ImGui::BulletText("Fixed World Position: Any controller can drag (HMD-only mode) or attached controller only (Both modes)");
-				ImGui::BulletText("HMD Relative: Any controller can drag (HMD-only mode) or attached controller only (Both modes)");
-				ImGui::BulletText("Controller Attached: Only the opposite hand can drag the controller overlay");
-				ImGui::Spacing();
-				ImGui::TextWrapped("Depth Adjustment (Grip + Thumbstick):");
-				ImGui::BulletText("While gripping to drag, use the thumbstick on the same hand to adjust depth");
-				ImGui::BulletText("Thumbstick forward: Push overlay farther away");
-				ImGui::BulletText("Thumbstick back: Pull overlay closer");
-			}
-			ImGui::Checkbox("Enable drag to reposition overlays", &settings.EnableDragToReposition);
-			ImGui::BeginDisabled(!settings.EnableDragToReposition);
-			ImGui::ColorEdit4("Drag Highlight Color", settings.dragHighlightColor.data());
-			ImGui::EndDisabled();
-			if (auto _tt = Util::HoverTooltipWrapper()) {
-				ImGui::Text("Color used to highlight draggable overlays in VR.");
-			}
-		}
-	}
-	void DrawKeyBindings()
-	{
-		auto& vr = globals::features::vr;
-		auto& settings = vr.settings;
-
-		// Combo Settings
-		if (ImGui::CollapsingHeader("Combo Settings", ImGuiTreeNodeFlags_DefaultOpen)) {
-			ImGui::SliderFloat("Combo Timeout", &settings.comboTimeout, 1.0f, 10.0f, "%.1f seconds");
-			if (auto _tt = Util::HoverTooltipWrapper()) {
-				ImGui::Text("Time limit for recording button combinations.");
-			}
-		}
-		ImGui::Separator();
-		// Combo box for selecting which combo to record
-		const char* comboTypes[] = {
-			"Open Community Shaders Menu",
-			"Close Community Shaders Menu",
-			"Open VR Overlay",
-			"Close VR Overlay"
-		};
-		static int selectedComboIndex = 0;
-		ImGui::Text("Select Combo to Record:");
-		ImGui::SameLine();
-		if (ImGui::Combo("##ComboSelector", &selectedComboIndex, comboTypes, IM_ARRAYSIZE(comboTypes))) {
-			// Reset recording state when changing selection
-			vr.isCapturingCombo = false;
-			vr.currentComboType = VR::ComboType::None;
-			vr.recordedCombo.clear();
-		}
-		if (ImGui::Button("Record Selected Combo")) {
-			// Start recording the selected combo
-			vr.isCapturingCombo = true;
-			vr.currentComboType = static_cast<VR::ComboType>(selectedComboIndex + 1);
-			vr.currentComboName = comboTypes[selectedComboIndex];
-			vr.recordedCombo.clear();
-			vr.comboStartTime = Util::GetNowSecs();
-			vr.recordingButtonControllers.clear();
-		}
-		ImGui::SameLine();
-		if (ImGui::SmallButton("Clear")) {
-			// Clear the selected combo
-			switch (selectedComboIndex) {
-			case 0:
-				settings.VRMenuOpenKeys.clear();
-				break;
-			case 1:
-				settings.VRMenuCloseKeys.clear();
-				break;
-			case 2:
-				settings.VROverlayOpenKeys.clear();
-				break;
-			case 3:
-				settings.VROverlayCloseKeys.clear();
-				break;
-			}
-		}
-		if (auto _tt = Util::HoverTooltipWrapper()) {
-			ImGui::Text("Click to start recording a new button combination for the selected action.");
-		}
-		ImGui::Spacing();
-		ImGui::Separator();
-		ImGui::Spacing();
-		// Table for displaying current key bindings
-		if (ImGui::BeginTable("##VRBindingsTable", 3, ImGuiTableFlags_Borders | ImGuiTableFlags_RowBg | ImGuiTableFlags_SizingStretchProp)) {
-			ImGui::TableSetupColumn("Action");
-			ImGui::TableSetupColumn("Current Binding");
-			ImGui::TableSetupColumn("Description");
-			ImGui::TableHeadersRow();
-			// Define VR key binding configurations
-			struct VRKeyBindingConfig
-			{
-				const char* label;
-				std::vector<InputCombo>& combos;
-				const char* description;
-				const char* controllerRequirement;
-			};
-			std::vector<VRKeyBindingConfig> keyBindingConfigs = {
-				{ "Open Community Shaders Menu", settings.VRMenuOpenKeys, "Button combination to open the Community Shaders menu", "Primary" },
-				{ "Close Community Shaders Menu", settings.VRMenuCloseKeys, "Button combination to close the Community Shaders menu", "Both" },
-				{ "Open VR Overlay", settings.VROverlayOpenKeys, "Button combination to open the VR overlay", "Primary" },
-				{ "Close VR Overlay", settings.VROverlayCloseKeys, "Button combination to close the VR overlay", "Secondary" }
-			};
-			for (size_t row = 0; row < keyBindingConfigs.size(); ++row) {
-				const auto& config = keyBindingConfigs[row];
-				ImGui::TableNextRow();
-				// Highlight the selected row
-				if (row == static_cast<size_t>(selectedComboIndex)) {
-					ImU32 highlight = ImGui::GetColorU32(ImVec4(1.0f, 1.0f, 0.0f, 0.15f));
-					ImGui::TableSetBgColor(ImGuiTableBgTarget_RowBg0, highlight);
-				}
-				// Make row selectable
-				ImGui::TableSetColumnIndex(0);
-				char selectableId[64];
-				snprintf(selectableId, sizeof(selectableId), "##combo_row_%zu", row);
-				bool rowSelected = (row == static_cast<size_t>(selectedComboIndex));
-				if (ImGui::Selectable(selectableId, rowSelected, ImGuiSelectableFlags_SpanAllColumns | ImGuiSelectableFlags_AllowItemOverlap, ImVec2(0, 0))) {
-					selectedComboIndex = static_cast<int>(row);
-				}
-				ImGui::SameLine(0, 0);
-				ImGui::Text("%s", config.label);
-				// Current Binding column
-				ImGui::TableSetColumnIndex(1);
-				Util::DrawButtonCombo(config.combos, false);
-				// Description column
-				ImGui::TableSetColumnIndex(2);
-				ImGui::Text("%s", config.description);
-			}
-			ImGui::EndTable();
-		}
-		ImGui::Spacing();
-		// Reset to defaults button
-		if (ImGui::Button("Reset to Defaults")) {
-			// Use InputCombo structure for cleaner defaults
-			settings.VRMenuOpenKeys = {
-				InputCombo::Primary(static_cast<uint32_t>(RE::BSOpenVRControllerDevice::Keys::kXA)),
-				InputCombo::Primary(static_cast<uint32_t>(RE::BSOpenVRControllerDevice::Keys::kBY))
-			};
-			settings.VRMenuCloseKeys = {
-				InputCombo::Both(static_cast<uint32_t>(RE::BSOpenVRControllerDevice::Keys::kGrip))
-			};
-			settings.VROverlayOpenKeys = {
-				InputCombo::Secondary(static_cast<uint32_t>(RE::BSOpenVRControllerDevice::Keys::kJoystickTrigger))
-			};
-			settings.VROverlayCloseKeys = {
-				InputCombo::Primary(static_cast<uint32_t>(RE::BSOpenVRControllerDevice::Keys::kJoystickTrigger))
-			};
-		}
-		if (auto _tt = Util::HoverTooltipWrapper()) {
-			ImGui::Text("Reset all VR key bindings to their default values.");
-		}
-	}
-	void DrawDebugSection()
-	{
-		auto& vr = globals::features::vr;
-		auto& settings = vr.settings;
-		auto menu = globals::menu;
-
-		// OpenVR Version Information
-		if (ImGui::CollapsingHeader("OpenVR Information", ImGuiTreeNodeFlags_DefaultOpen)) {
-			auto& info = vr.openVRInfo;
-			if (info.isAvailable) {
-				if (vr.IsOpenVRCompatible()) {
-					ImGui::Text("OpenVR System: Active & Compatible");
-				} else {
-					ImGui::TextColored(ImVec4(1.0f, 0.5f, 0.5f, 1.0f), "OpenVR System: Active but INCOMPATIBLE");
-					ImGui::TextColored(ImVec4(1.0f, 0.5f, 0.5f, 1.0f), "VR overlay menus disabled.");
-				}
-
-				ImGui::Text("Runtime: %s", VRDetection::RuntimeTypeToString(info.runtimeType));
-				ImGui::Text("DLL Path: %s", info.dllPath.c_str());
-				ImGui::Text("DLL Version: %s", info.version.c_str());
-				ImGui::Text("DLL Size: %llu bytes", info.fileSize);
-				ImGui::Text("Modified: %s", info.modificationTime.c_str());
-
-				ImGui::Separator();
-				ImGui::Text("Detection Method:");
-				ImGui::Text("  Interface Probing: %s", info.probingSucceeded ? "Passed" : "Failed");
-				ImGui::Text("    IVROverlay_016: %s", info.hasOverlayInterface ? "OK" : "Missing");
-				ImGui::Text("    IVRSystem_017: %s", info.hasSystemInterface ? "OK" : "Missing");
-				ImGui::Text("    IVRCompositor_021: %s", info.hasCompositorInterface ? "OK" : "Missing");
-				ImGui::TextColored(ImVec4(0.3f, 1.0f, 0.3f, 1.0f), "  Rendering: In-scene overlay (submit hook)");
-
-			} else {
-				ImGui::Text("OpenVR system not available");
-			}
-		}
-
-		// Controller Diagnostics Section
-		if (ImGui::CollapsingHeader("Controller Diagnostics", ImGuiTreeNodeFlags_DefaultOpen)) {
-			if (ImGui::Checkbox("Test Mode: Disable controller menu input (except scroll controller and triggers)", &settings.VRMenuControllerDiagnosticsTestMode)) {
-				ImGui::SetScrollHereY(0.0f);  // Scroll to top of the window when toggled
-			}
-			ImGui::SeparatorText("Button State");
-			double nowSecs = Util::GetNowSecs();
-			// Get highlight color from theme
-			ImVec4 highlightColor = menu->GetTheme().StatusPalette.InfoColor;
-			ImU32 highlightColorU32 = ImGui::ColorConvertFloat4ToU32(highlightColor);
-
-			// Determine display order based on handedness
-			bool isLeftHanded = vr.lastKnownLeftHandedMode;  // Use cached handedness
-
-			if (ImGui::BeginTable("vr_input_state_table", 7, ImGuiTableFlags_Borders | ImGuiTableFlags_RowBg)) {
-				ImGui::TableSetupColumn("Button");
-				if (isLeftHanded) {
-					// Left-handed: Primary (left hand) on left, Secondary (right hand) on right
-					ImGui::TableSetupColumn("Primary State");
-					ImGui::TableSetupColumn("Primary Held (s)");
-					ImGui::TableSetupColumn("Primary Type");
-					ImGui::TableSetupColumn("Secondary State");
-					ImGui::TableSetupColumn("Secondary Held (s)");
-					ImGui::TableSetupColumn("Secondary Type");
-				} else {
-					// Right-handed: Secondary (left hand) on left, Primary (right hand) on right
-					ImGui::TableSetupColumn("Secondary State");
-					ImGui::TableSetupColumn("Secondary Held (s)");
-					ImGui::TableSetupColumn("Secondary Type");
-					ImGui::TableSetupColumn("Primary State");
-					ImGui::TableSetupColumn("Primary Held (s)");
-					ImGui::TableSetupColumn("Primary Type");
-				}
-				ImGui::TableHeadersRow();
-				// Helper for button type text
-				auto DrawButtonType = [](const RE::ButtonState& state) {
-					if (!state.isPressed) {
-						if (state.IsClick())
-							ImGui::TextUnformatted("Click");
-						else if (state.IsHold())
-							ImGui::TextUnformatted("Hold");
-						else
-							ImGui::TextUnformatted("-");
-					} else {
-						ImGui::TextUnformatted("Held");
-					}
-				};
-				// Helper for printing a row with left/right cell highlight
-				auto printRow = [&](const char* label, const RE::ButtonState& left, const RE::ButtonState& right) {
-					ImGui::TableNextRow();
-					ImGui::TableSetColumnIndex(0);
-					ImGui::TextUnformatted(label);
-					ImGui::TableSetColumnIndex(1);
-					if (left.isPressed)
-						ImGui::TableSetBgColor(ImGuiTableBgTarget_CellBg, highlightColorU32);
-					ImGui::TextUnformatted(left.isPressed ? "Pressed" : "Released");
-					ImGui::TableSetColumnIndex(2);
-					if (left.isPressed)
-						ImGui::TableSetBgColor(ImGuiTableBgTarget_CellBg, highlightColorU32);
-					ImGui::Text("%.2f", left.GetCurrentHeldTime(nowSecs));
-					ImGui::TableSetColumnIndex(3);
-					if (left.isPressed)
-						ImGui::TableSetBgColor(ImGuiTableBgTarget_CellBg, highlightColorU32);
-					DrawButtonType(left);
-					ImGui::TableSetColumnIndex(4);
-					if (right.isPressed)
-						ImGui::TableSetBgColor(ImGuiTableBgTarget_CellBg, highlightColorU32);
-					ImGui::TextUnformatted(right.isPressed ? "Pressed" : "Released");
-					ImGui::TableSetColumnIndex(5);
-					if (right.isPressed)
-						ImGui::TableSetBgColor(ImGuiTableBgTarget_CellBg, highlightColorU32);
-					ImGui::Text("%.2f", right.GetCurrentHeldTime(nowSecs));
-					ImGui::TableSetColumnIndex(6);
-					if (right.isPressed)
-						ImGui::TableSetBgColor(ImGuiTableBgTarget_CellBg, highlightColorU32);
-					DrawButtonType(right);
-				};
-
-				// Helper to determine the correct order for display based on handedness
-				auto printRowWithHandedness = [&](const char* label, auto key) {
-					auto& primary = vr.primaryControllerState[key];
-					auto& secondary = vr.secondaryControllerState[key];
-					if (isLeftHanded) {
-						// Left-handed: Primary (left hand) on left, Secondary (right hand) on right
-						printRow(label, primary, secondary);
-					} else {
-						// Right-handed: Secondary (left hand) on left, Primary (right hand) on right
-						printRow(label, secondary, primary);
-					}
-				};
-
-				printRowWithHandedness("Trigger", RE::BSOpenVRControllerDevice::Keys::kTrigger);
-				printRowWithHandedness("Grip", RE::BSOpenVRControllerDevice::Keys::kGrip);
-				printRowWithHandedness("GripAlt", RE::BSOpenVRControllerDevice::Keys::kGripAlt);
-				printRowWithHandedness("Stick Click", RE::BSOpenVRControllerDevice::Keys::kJoystickTrigger);
-				printRowWithHandedness("Touchpad Click", RE::BSOpenVRControllerDevice::Keys::kTouchpadClick);
-				printRowWithHandedness("Touchpad Alt", RE::BSOpenVRControllerDevice::Keys::kTouchpadAlt);
-				printRowWithHandedness("B/Y", RE::BSOpenVRControllerDevice::Keys::kBY);
-				printRowWithHandedness("A/X", RE::BSOpenVRControllerDevice::Keys::kXA);
-				ImGui::EndTable();
-			}
-			ImGui::SeparatorText("VR Thumbstick State");
-			// Helper to draw a thumbstick quadrant visualization (returns ImVec2 for label alignment)
-			auto DrawThumbstickPad = [&](float x, float y, ImU32 highlightCol) -> ImVec2 {
-				ImVec2 padSize = ImVec2(80, 80);
-				ImVec2 cursor = ImGui::GetCursorScreenPos();
-				ImDrawList* drawList = ImGui::GetWindowDrawList();
-				ImVec2 center = ImVec2(cursor.x + padSize.x / 2, cursor.y + padSize.y / 2);
-				float radius = padSize.x / 2 - 4;
-				ImU32 borderCol = ImGui::GetColorU32(ImGuiCol_Border);
-				ImU32 axisCol = ImGui::GetColorU32(ImGuiCol_TextDisabled);
-				ImU32 dotCol = ImGui::GetColorU32(ImGuiCol_Text);
-				// Draw background
-				drawList->AddRectFilled(cursor, ImVec2(cursor.x + padSize.x, cursor.y + padSize.y), ImGui::GetColorU32(ImGuiCol_FrameBg));
-				// Draw border
-				drawList->AddRect(cursor, ImVec2(cursor.x + padSize.x, cursor.y + padSize.y), borderCol, 4.0f, 0, 2.0f);
-				// Draw axes
-				drawList->AddLine(ImVec2(center.x, cursor.y + 4), ImVec2(center.x, cursor.y + padSize.y - 4), axisCol, 1.0f);
-				drawList->AddLine(ImVec2(cursor.x + 4, center.y), ImVec2(cursor.x + padSize.x - 4, center.y), axisCol, 1.0f);
-				// Determine quadrant
-				int quad = 0;
-				if (x > 0 && y > 0)
-					quad = 1;  // top-right
-				else if (x < 0 && y > 0)
-					quad = 2;  // top-left
-				else if (x < 0 && y < 0)
-					quad = 3;  // bottom-left
-				else if (x > 0 && y < 0)
-					quad = 4;  // bottom-right
-				// Highlight quadrant
-				if (quad != 0) {
-					ImVec2 q0 = center;
-					ImVec2 q1 = center;
-					ImVec2 q2 = center;
-					ImVec2 q3 = center;
-					if (quad == 1) {  // top-right
-						q1.x += radius;
-						q1.y -= radius;
-						q2.x += radius;
-						q2.y += 0;
-						q3.x += 0;
-						q3.y -= radius;
-					} else if (quad == 2) {  // top-left
-						q1.x -= radius;
-						q1.y -= radius;
-						q2.x -= radius;
-						q2.y += 0;
-						q3.x += 0;
-						q3.y -= radius;
-					} else if (quad == 3) {  // bottom-left
-						q1.x -= radius;
-						q1.y += radius;
-						q2.x -= radius;
-						q2.y += 0;
-						q3.x += 0;
-						q3.y += radius;
-					} else if (quad == 4) {  // bottom-right
-						q1.x += radius;
-						q1.y += radius;
-						q2.x += radius;
-						q2.y += 0;
-						q3.x += 0;
-						q3.y += radius;
-					}
-					ImVec2 poly[4] = { center, q1, q2, q3 };
-					drawList->AddConvexPolyFilled(poly, 4, highlightCol);
-				}
-				// Draw stick position dot
-				ImVec2 dot = ImVec2(center.x + x * radius, center.y - y * radius);
-				drawList->AddCircleFilled(dot, 5.0f, dotCol);
-				// Return size for label alignment
-				return padSize;
-			};
-			ImU32 highlightCol = ImGui::ColorConvertFloat4ToU32(menu->GetTheme().StatusPalette.InfoColor);
-			if (ImGui::BeginTable("##VRThumbstickTable", 2, ImGuiTableFlags_Borders | ImGuiTableFlags_RowBg | ImGuiTableFlags_SizingFixedFit)) {
-				if (isLeftHanded) {
-					// Left-handed: Primary (left hand) on left, Secondary (right hand) on right
-					ImGui::TableSetupColumn("Primary Controller", ImGuiTableColumnFlags_WidthFixed, 200.0f);
-					ImGui::TableSetupColumn("Secondary Controller", ImGuiTableColumnFlags_WidthFixed, 200.0f);
-				} else {
-					// Right-handed: Secondary (left hand) on left, Primary (right hand) on right
-					ImGui::TableSetupColumn("Secondary Controller", ImGuiTableColumnFlags_WidthFixed, 200.0f);
-					ImGui::TableSetupColumn("Primary Controller", ImGuiTableColumnFlags_WidthFixed, 200.0f);
-				}
-				ImGui::TableHeadersRow();
-
-				// Left column content
-				ImGui::TableSetColumnIndex(0);
-				ImGui::BeginGroup();
-				if (isLeftHanded) {
-					// Left-handed: Show primary controller in left column
-					ImVec2 padSizeL = DrawThumbstickPad(vr.primaryControllerState.thumbsticks[static_cast<size_t>(RE::ControllerRole::Primary)].x, vr.primaryControllerState.thumbsticks[static_cast<size_t>(RE::ControllerRole::Primary)].y, highlightCol);
-					ImGui::Dummy(padSizeL);
-					ImGui::SetNextItemWidth(160.0f);
-					ImGui::SetCursorPosY(ImGui::GetCursorPosY() - ImGui::GetTextLineHeight());
-					ImGui::Text("X: %+1.3f  Y: %+1.3f  [%s]", vr.primaryControllerState.thumbsticks[static_cast<size_t>(RE::ControllerRole::Primary)].x, vr.primaryControllerState.thumbsticks[static_cast<size_t>(RE::ControllerRole::Primary)].y, RE::GetQuadrantName(vr.primaryControllerState.thumbsticks[static_cast<size_t>(RE::ControllerRole::Primary)].x, vr.primaryControllerState.thumbsticks[static_cast<size_t>(RE::ControllerRole::Primary)].y));
-				} else {
-					// Right-handed: Show secondary controller in left column
-					ImVec2 padSizeL = DrawThumbstickPad(vr.secondaryControllerState.thumbsticks[static_cast<size_t>(RE::ControllerRole::Secondary)].x, vr.secondaryControllerState.thumbsticks[static_cast<size_t>(RE::ControllerRole::Secondary)].y, highlightCol);
-					ImGui::Dummy(padSizeL);
-					ImGui::SetNextItemWidth(160.0f);
-					ImGui::SetCursorPosY(ImGui::GetCursorPosY() - ImGui::GetTextLineHeight());
-					ImGui::Text("X: %+1.3f  Y: %+1.3f  [%s]", vr.secondaryControllerState.thumbsticks[static_cast<size_t>(RE::ControllerRole::Secondary)].x, vr.secondaryControllerState.thumbsticks[static_cast<size_t>(RE::ControllerRole::Secondary)].y, RE::GetQuadrantName(vr.secondaryControllerState.thumbsticks[static_cast<size_t>(RE::ControllerRole::Secondary)].x, vr.secondaryControllerState.thumbsticks[static_cast<size_t>(RE::ControllerRole::Secondary)].y));
-				}
-				ImGui::EndGroup();
-
-				// Right column content
-				ImGui::TableSetColumnIndex(1);
-				ImGui::BeginGroup();
-				if (isLeftHanded) {
-					// Left-handed: Show secondary controller in right column
-					ImVec2 padSizeR = DrawThumbstickPad(vr.secondaryControllerState.thumbsticks[static_cast<size_t>(RE::ControllerRole::Secondary)].x, vr.secondaryControllerState.thumbsticks[static_cast<size_t>(RE::ControllerRole::Secondary)].y, highlightCol);
-					ImGui::Dummy(padSizeR);
-					ImGui::SetNextItemWidth(160.0f);
-					ImGui::SetCursorPosY(ImGui::GetCursorPosY() - ImGui::GetTextLineHeight());
-					ImGui::Text("X: %+1.3f  Y: %+1.3f  [%s]", vr.secondaryControllerState.thumbsticks[static_cast<size_t>(RE::ControllerRole::Secondary)].x, vr.secondaryControllerState.thumbsticks[static_cast<size_t>(RE::ControllerRole::Secondary)].y, RE::GetQuadrantName(vr.secondaryControllerState.thumbsticks[static_cast<size_t>(RE::ControllerRole::Secondary)].x, vr.secondaryControllerState.thumbsticks[static_cast<size_t>(RE::ControllerRole::Secondary)].y));
-				} else {
-					// Right-handed: Show primary controller in right column
-					ImVec2 padSizeR = DrawThumbstickPad(vr.primaryControllerState.thumbsticks[static_cast<size_t>(RE::ControllerRole::Primary)].x, vr.primaryControllerState.thumbsticks[static_cast<size_t>(RE::ControllerRole::Primary)].y, highlightCol);
-					ImGui::Dummy(padSizeR);
-					ImGui::SetNextItemWidth(160.0f);
-					ImGui::SetCursorPosY(ImGui::GetCursorPosY() - ImGui::GetTextLineHeight());
-					ImGui::Text("X: %+1.3f  Y: %+1.3f  [%s]", vr.primaryControllerState.thumbsticks[static_cast<size_t>(RE::ControllerRole::Primary)].x, vr.primaryControllerState.thumbsticks[static_cast<size_t>(RE::ControllerRole::Primary)].y, RE::GetQuadrantName(vr.primaryControllerState.thumbsticks[static_cast<size_t>(RE::ControllerRole::Primary)].x, vr.primaryControllerState.thumbsticks[static_cast<size_t>(RE::ControllerRole::Primary)].y));
-				}
-				ImGui::EndGroup();
-				ImGui::EndTable();
-			}
-			ImGui::SeparatorText("Recent VR Controller Events");
-			ImGui::TextDisabled("Note: For thumbstick events, KeyCode/Value columns show X/Y floats.");
-			if (ImGui::BeginTable("eventlog", 6, ImGuiTableFlags_Borders | ImGuiTableFlags_RowBg | ImGuiTableFlags_SizingFixedFit)) {
-				ImGui::TableSetupColumn("Device", ImGuiTableColumnFlags_WidthFixed, 60.0f);
-				ImGui::TableSetupColumn("KeyCode/X", ImGuiTableColumnFlags_WidthFixed, 80.0f);
-				ImGui::TableSetupColumn("Value/Y", ImGuiTableColumnFlags_WidthFixed, 80.0f);
-				ImGui::TableSetupColumn("Pressed", ImGuiTableColumnFlags_WidthFixed, 70.0f);
-				ImGui::TableSetupColumn("Known Mapping", ImGuiTableColumnFlags_WidthFixed, 120.0f);
-				ImGui::TableSetupColumn("Event Type", ImGuiTableColumnFlags_WidthFixed, 120.0f);
-				ImGui::TableHeadersRow();
-				for (const auto& e : vr.vrControllerEventLog) {
-					ImGui::TableNextRow();
-					ImGui::TableSetColumnIndex(0);
-					ImGui::Text("%d", e.device);
-					ImGui::TableSetColumnIndex(1);
-					if (e.heldSource == "thumbstick") {
-						ImGui::Text("%.3f", e.thumbstickX);
-					} else {
-						ImGui::Text("%d", e.keyCode);
-					}
-					ImGui::TableSetColumnIndex(2);
-					if (e.heldSource == "thumbstick") {
-						ImGui::Text("%.3f", e.thumbstickY);
-					} else {
-						ImGui::Text("%d", e.value);
-					}
-					ImGui::TableSetColumnIndex(3);
-					ImGui::Text("%s", e.pressed ? "Pressed" : "Released");
-					ImGui::TableSetColumnIndex(4);
-					if (e.heldSource == "thumbstick") {
-						ImGui::TextUnformatted(e.controllerRole.c_str());
-					} else {
-						ImGui::TextUnformatted(RE::GetOpenVRButtonName(e.keyCode));
-					}
-					ImGui::TableSetColumnIndex(5);
-					if (e.heldSource == "thumbstick") {
-						ImGui::TextUnformatted("-");
-					} else {
-						// Show click/hold for release events if available
-						if (!e.pressed) {
-							if (e.heldTime > 0.0) {
-								if (e.heldTime < 0.5) {
-									ImGui::Text("Click (%.2fs)", e.heldTime);
-								} else {
-									ImGui::Text("Hold (%.2fs)", e.heldTime);
-								}
-							} else {
-								ImGui::Text("Release");
-							}
-						} else if (e.pressed) {
-							if (e.heldTime > 0.0) {
-								ImGui::Text("Held for %.2fs", e.heldTime);
-							} else {
-								ImGui::Text("Press");
-							}
-						}
-					}
-				}
-				ImGui::EndTable();
-			}
-
-			// Wand Pointing Diagnostics
-			ImGui::SeparatorText("Wand Pointing State");
-			if (ImGui::BeginTable("##WandPointingState", 2, ImGuiTableFlags_Borders | ImGuiTableFlags_RowBg)) {
-				ImGui::TableSetupColumn("Property", ImGuiTableColumnFlags_WidthFixed, 200.0f);
-				ImGui::TableSetupColumn("Value", ImGuiTableColumnFlags_WidthStretch);
-				ImGui::TableHeadersRow();
-
-				ImGui::TableNextRow();
-				ImGui::TableSetColumnIndex(0);
-				ImGui::Text("Wand Pointing Enabled");
-				ImGui::TableSetColumnIndex(1);
-				ImGui::Text("%s", settings.EnableWandPointing ? "Yes" : "No");
-
-				ImGui::TableNextRow();
-				ImGui::TableSetColumnIndex(0);
-				ImGui::Text("Intersecting Overlay");
-				ImGui::TableSetColumnIndex(1);
-				if (vr.wandState.isIntersecting) {
-					ImGui::TextColored(menu->GetTheme().StatusPalette.InfoColor, "YES");
-				} else {
-					ImGui::Text("No");
-				}
-
-				ImGui::TableNextRow();
-				ImGui::TableSetColumnIndex(0);
-				ImGui::Text("UV Coordinates");
-				ImGui::TableSetColumnIndex(1);
-				ImGui::Text("(%.3f, %.3f)", vr.wandState.uvCoordinates.x, vr.wandState.uvCoordinates.y);
-
-				ImGui::TableNextRow();
-				ImGui::TableSetColumnIndex(0);
-				ImGui::Text("Controller Index");
-				ImGui::TableSetColumnIndex(1);
-				ImGui::Text("%u", vr.wandState.controllerIndex);
-
-				ImGui::TableNextRow();
-				ImGui::TableSetColumnIndex(0);
-				ImGui::Text("Ray Origin");
-				ImGui::TableSetColumnIndex(1);
-				ImGui::Text("(%.2f, %.2f, %.2f)", vr.wandState.rayOrigin.x, vr.wandState.rayOrigin.y, vr.wandState.rayOrigin.z);
-
-				ImGui::TableNextRow();
-				ImGui::TableSetColumnIndex(0);
-				ImGui::Text("Ray Direction");
-				ImGui::TableSetColumnIndex(1);
-				ImGui::Text("(%.2f, %.2f, %.2f)", vr.wandState.rayDirection.x, vr.wandState.rayDirection.y, vr.wandState.rayDirection.z);
-
-				ImGui::EndTable();
-			}
-		}
-
-		// Debugging addresses for copy/paste
-		if (ImGui::CollapsingHeader("OpenVR Addresses")) {
-			auto openvr = RE::BSOpenVR::GetSingleton();
-			auto overlay = openvr ? RE::BSOpenVR::GetIVROverlayFromContext(&openvr->vrContext) : nullptr;
-			auto vrSystem = openvr ? openvr->vrSystem : nullptr;
-			ADDRESS_NODE(openvr)
-			ADDRESS_NODE(overlay)
-			ADDRESS_NODE(vrSystem)
-		}
-	}
-}  // namespace
-
-//=============================================================================
-// VR-SPECIFIC PUBLIC API
-//=============================================================================
-// PRIVATE IMPLEMENTATION
+// OVERLAY SUBMIT AND DEPTH BUFFER CULLING
 //=============================================================================
 
 void VR::RecreateOverlayTexturesIfNeeded()
 {
-	// Only need the main menu texture (used for all attach modes via submit hook)
 	Util::CreateOverlayTextureAndRTV(globals::d3d::device, Config::kOverlayWidth, Config::kOverlayHeight, menuTexture.put(), menuRTV.put());
 }
 
-// Wraps IVROverlay calls that may trigger OpenComposite's abort (int3) inside SEH.
 void VR::SubmitOverlayFrame()
 {
-	// The submit hook is the primary rendering method -- install it unconditionally
 	InstallSubmitHook();
 
 	RE::BSOpenVR* openvr = RE::BSOpenVR::GetSingleton();
@@ -1332,7 +175,6 @@ void VR::SubmitOverlayFrame()
 		UpdateFixedWorldPositioning();
 		UpdateOverlayDrag();
 
-		// Render ImGui to the overlay texture
 		ID3D11RenderTargetView* oldRTV = nullptr;
 		globals::d3d::context->OMGetRenderTargets(1, &oldRTV, nullptr);
 		ID3D11RenderTargetView* menuRTVPtr = menuRTV.get();
@@ -1343,7 +185,6 @@ void VR::SubmitOverlayFrame()
 		ImGui_ImplDX11_RenderDrawData(ImGui::GetDrawData());
 		globals::d3d::context->OMSetRenderTargets(1, &oldRTV, nullptr);
 
-		// Apply highlight tint if being dragged (any mode, since both overlays share this texture)
 		bool beingDragged = settings.EnableDragToReposition && overlayDragState.dragging;
 		Util::ApplyHighlightTintToTexture(menuTexture.get(), beingDragged, settings.dragHighlightColor);
 
@@ -1352,14 +193,11 @@ void VR::SubmitOverlayFrame()
 	}
 }
 
-// Helper to centralize VR depth buffer culling logic, reducing duplication between DataLoaded and EarlyPrepass.
 void VR::UpdateDepthBufferCulling(bool desired, const FeatureConstraints::SettingId& settingId)
 {
-	// Check if any feature is constraining this setting
 	auto constraint = FeatureConstraints::GetConstraints(settingId);
 
 	if (constraint.isConstrained) {
-		// Use std::get_if to safely extract bool value and avoid std::bad_variant_access
 		if (auto* forcedValuePtr = std::get_if<bool>(&constraint.forcedValue)) {
 			bool forcedValue = *forcedValuePtr;
 			if (gDepthBufferCulling && *gDepthBufferCulling != forcedValue) {
@@ -1372,906 +210,12 @@ void VR::UpdateDepthBufferCulling(bool desired, const FeatureConstraints::Settin
 				}
 			}
 		} else {
-			// Constraint has non-bool value type - log warning and skip
 			logger::warn("VR::UpdateDepthBufferCulling: Constraint on {} has non-bool forced value, ignoring", settingId.settingPath);
 		}
 	} else {
 		if (gDepthBufferCulling && *gDepthBufferCulling != desired) {
 			*gDepthBufferCulling = desired;
 			logger::info("VR depth buffer culling set to {}", desired);
-		}
-	}
-}
-
-// Handles overlay/menu open/close logic based on controller input state
-void VR::UpdateOverlayMenuStateFromInput()
-{
-	// Disable menu interactions during combo recording
-	if (this->isCapturingCombo) {
-		return;
-	}
-
-	bool& isEnabled = globals::menu->IsEnabled;
-	bool& overlayEnabled = globals::menu->overlayVisible;
-	bool& testMode = settings.VRMenuControllerDiagnosticsTestMode;
-
-	// Auto-disable test mode if user leaves VR section or closes menu
-	if (testMode) {
-		// Check if we're still in the VR section or if menu is closed
-		if (!isEnabled) {
-			settings.VRMenuControllerDiagnosticsTestMode = false;
-			return;
-		}
-		// In test mode, only allow basic input processing
-		return;
-	}
-
-	// Compute whether the game's UI menus we care about are open. Do this early so
-	// downstream logic can reuse the result and we only check the UI once.
-	bool uiMenusOpen = globals::game::ui &&
-	                   (globals::game::ui->IsMenuOpen(RE::MainMenu::MENU_NAME) || globals::game::ui->IsMenuOpen(RE::TweenMenu::MENU_NAME));
-
-	// Valid menu state means either one of those UI menus is open, or our menu/overlay is
-	// enabled (but only if the game's UI system is present).
-	bool inValidMenuState = uiMenusOpen || (globals::game::ui && (isEnabled || overlayEnabled));
-
-	if (!inValidMenuState)
-		return;
-
-	// Define menu state mappings. The `allowWhenUIMenusClosed` flag controls whether
-	// a mapping is allowed to run when our menu is enabled but the game's UI menus
-	// are not reported open. This prevents 'open' controls from firing in that state
-	// while still allowing 'close' actions.
-	struct MenuStateMapping
-	{
-		std::function<bool()> condition;
-		std::function<void()> action;
-		bool allowWhenUIMenusClosed = false;
-	};
-
-	// Generic combo checking function - makes the system truly extensible
-	auto CheckCombo = [&](const std::vector<ButtonCombo>& combos) -> bool {
-		if (combos.empty())
-			return false;
-
-		// Check all configured buttons in the combo
-		for (size_t i = 0; i < combos.size(); ++i) {
-			const auto& combo = combos[i];
-
-			bool buttonPressed = false;
-
-			switch (combo.GetDevice()) {
-			case ControllerDevice::Both:
-				buttonPressed = primaryControllerState[combo.GetKey()].isPressed &&
-				                secondaryControllerState[combo.GetKey()].isPressed;
-				break;
-			case ControllerDevice::Primary:
-				buttonPressed = primaryControllerState[combo.GetKey()].isPressed;
-				break;
-			case ControllerDevice::Secondary:
-				buttonPressed = secondaryControllerState[combo.GetKey()].isPressed;
-				break;
-			}
-
-			if (!buttonPressed) {
-				return false;
-			}
-		}
-
-		return true;
-	};
-
-	// Define the menu state mappings with extensible lambda array
-	std::vector<MenuStateMapping> mappings = {
-		// Open Community Shaders menu when closed
-		{ [&]() {
-			 return CheckCombo(settings.VRMenuOpenKeys) && !isEnabled;
-		 },
-			[&]() { isEnabled = true; } },
-
-		// Close Community Shaders menu when open
-		{ [&]() {
-			 return CheckCombo(settings.VRMenuCloseKeys) && isEnabled;
-		 },
-			[&]() {
-				isEnabled = false;
-				// Reset drag state to clear highlight when menu closes
-				overlayDragState.dragging = false;
-			},
-			true },
-
-		// Open VR overlay when closed (only when CS menu is open)
-		{ [&]() {
-			 return CheckCombo(settings.VROverlayOpenKeys) && !overlayEnabled && isEnabled;
-		 },
-			[&]() { overlayEnabled = true; } },
-
-		// Close VR overlay when open (only when CS menu is open)
-		{ [&]() {
-			 return CheckCombo(settings.VROverlayCloseKeys) && overlayEnabled && isEnabled;
-		 },
-			[&]() { overlayEnabled = false; } }
-	};
-
-	// Process mappings in order. If our menu is enabled but the game's UI menus
-	// are not open, only allow mappings explicitly marked with
-	// allowWhenUIMenusClosed (close actions).
-	bool onlyAllowClose = isEnabled && !uiMenusOpen;
-
-	for (const auto& mapping : mappings) {
-		if (onlyAllowClose && !mapping.allowWhenUIMenusClosed)
-			continue;
-
-		if (mapping.condition()) {
-			mapping.action();
-			break;  // Only execute one action per frame
-		}
-	}
-}
-
-void VR::ProcessVREvents(std::vector<Menu::KeyEvent>& vrEvents)
-{
-	// Check for handedness changes and reset controller states if needed
-	bool currentLeftHandedMode = RE::BSOpenVRControllerDevice::IsLeftHandedMode();
-	static bool firstCall = true;
-	if (firstCall || currentLeftHandedMode != lastKnownLeftHandedMode) {
-		if (!firstCall) {
-			logger::debug("VR handedness changed: {} -> {}", lastKnownLeftHandedMode ? "Left" : "Right", currentLeftHandedMode ? "Left" : "Right");
-		}
-		firstCall = false;
-		lastKnownLeftHandedMode = currentLeftHandedMode;
-		// Reset controller states so they get repopulated with correct roles
-		primaryControllerState = {};
-		secondaryControllerState = {};
-	}
-
-	double nowSecs = Util::GetNowSecs();
-	for (auto& event : vrEvents) {
-		bool isPrimary = RE::BSOpenVRControllerDevice::IsPrimaryController(event.device);
-		bool isSecondary = RE::BSOpenVRControllerDevice::IsSecondaryController(event.device);
-		struct VRButtonDescriptor
-		{
-			const char* name;
-			bool (*isButton)(std::uint32_t);
-			std::uint32_t keyCode;
-		};
-		static const VRButtonDescriptor kVRButtons[] = {
-			{ "Grip", RE::BSOpenVRControllerDevice::IsGripButton, RE::BSOpenVRControllerDevice::Keys::kGrip },
-			{ "GripAlt", RE::BSOpenVRControllerDevice::IsGripButton, RE::BSOpenVRControllerDevice::Keys::kGripAlt },
-			{ "Trigger", RE::BSOpenVRControllerDevice::IsTriggerButton, RE::BSOpenVRControllerDevice::Keys::kTrigger },
-			{ "Stick Click", RE::BSOpenVRControllerDevice::IsStickClick, RE::BSOpenVRControllerDevice::Keys::kJoystickTrigger },
-			{ "Touchpad Click", RE::BSOpenVRControllerDevice::IsTouchpadClick, RE::BSOpenVRControllerDevice::Keys::kTouchpadClick },
-			{ "Touchpad Alt", RE::BSOpenVRControllerDevice::IsTouchpadClick, RE::BSOpenVRControllerDevice::Keys::kTouchpadAlt },
-			{ "A/X", RE::BSOpenVRControllerDevice::IsAButton, RE::BSOpenVRControllerDevice::Keys::kXA },
-			{ "B/Y", RE::BSOpenVRControllerDevice::IsBButton, RE::BSOpenVRControllerDevice::Keys::kBY },
-		};
-		for (const auto& desc : kVRButtons) {
-			if (desc.isButton(event.keyCode)) {
-				RE::ButtonState* state = isPrimary ? &primaryControllerState[desc.keyCode] : isSecondary ? &secondaryControllerState[desc.keyCode] :
-				                                                                                           nullptr;
-				if (state) {
-					state->OnEvent(event.IsPressed(), nowSecs);
-				}
-				break;
-			}
-		}
-		// Do not log events here; logging is handled in the event-specific handler
-		switch (event.eventType) {
-		case RE::INPUT_EVENT_TYPE::kButton:
-			ProcessVRButtonEvent(event);
-			break;
-		case RE::INPUT_EVENT_TYPE::kThumbstick:
-			UpdateControllerState(event);
-			break;
-		default:
-			break;
-		}
-	}
-}
-
-void VR::ProcessVRButtonEvent(const Menu::KeyEvent& event)
-{
-	// Disable menu interactions during combo recording
-	if (this->isCapturingCombo) {
-		return;
-	}
-
-	ImGuiIO& io = ImGui::GetIO();
-	(void)event;
-	bool isPrimary = RE::BSOpenVRControllerDevice::IsPrimaryController(event.device);
-	bool isSecondary = RE::BSOpenVRControllerDevice::IsSecondaryController(event.device);
-	bool& testMode = settings.VRMenuControllerDiagnosticsTestMode;
-	constexpr size_t kNumTriggerMappings = 1;
-
-	// Process button mappings for the current controller
-	if (isPrimary || isSecondary) {
-		// Define mappings for both controllers (only B/Y differs)
-		constexpr size_t kNumMappings = 6;
-		RE::ButtonMapping mappings[kNumMappings] = {
-			{ RE::BSOpenVRControllerDevice::Keys::kTrigger, ImGuiMouseButton_Left, false, ImGuiKey_None, false },
-			{ RE::BSOpenVRControllerDevice::Keys::kGrip, ImGuiMouseButton_Right, false, ImGuiKey_None, false },
-			{ RE::BSOpenVRControllerDevice::Keys::kTouchpadClick, ImGuiMouseButton_Middle, false, ImGuiKey_None, false },
-			{ RE::BSOpenVRControllerDevice::Keys::kJoystickTrigger, ImGuiMouseButton_Middle, false, ImGuiKey_None, false },
-			{ RE::BSOpenVRControllerDevice::Keys::kBY, -1, true, Util::Input::VirtualKeyToImGuiKey(VK_TAB), isSecondary },  // Shift+Tab for secondary
-			{ RE::BSOpenVRControllerDevice::Keys::kXA, -1, true, Util::Input::VirtualKeyToImGuiKey(VK_RETURN), false },
-		};
-
-		// Use separate state arrays for each controller
-		static bool prevPrimaryStates[kNumMappings] = {};
-		static bool prevSecondaryStates[kNumMappings] = {};
-		bool* prevStates = isPrimary ? prevPrimaryStates : prevSecondaryStates;
-
-		// Get the appropriate controller state
-		RE::InputDeviceState& controllerState = isPrimary ? primaryControllerState : secondaryControllerState;
-
-		size_t limit = testMode ? kNumTriggerMappings : kNumMappings;  // Only trigger mappings in test mode
-
-		for (size_t i = 0; i < limit; ++i) {
-			RE::ButtonState* state = &controllerState[mappings[i].keyCode];
-			bool curr = state ? state->isPressed : false;
-			if (curr != prevStates[i]) {
-				if (mappings[i].isKeyEvent) {
-					if (mappings[i].isShift)
-						io.AddKeyEvent(ImGuiMod_Shift, curr);
-					io.AddKeyEvent(static_cast<ImGuiKey>(mappings[i].key), curr);
-				} else {
-					io.AddMouseButtonEvent(mappings[i].logicalButton, curr);
-				}
-				prevStates[i] = curr;
-			}
-		}
-	}
-	// Log the button event after state is updated
-	VRControllerEventLog logEntry;
-	logEntry.device = static_cast<int>(event.device);
-	logEntry.keyCode = event.keyCode;
-	logEntry.value = static_cast<int>(event.value);
-	logEntry.pressed = event.IsPressed();
-	logEntry.heldTime = 0.0;
-	logEntry.heldSource = "button";
-	logEntry.thumbstickX = 0.0f;
-	logEntry.thumbstickY = 0.0f;
-	logEntry.controllerRole = isPrimary ? "Primary" : isSecondary ? "Secondary" :
-	                                                                "Unknown";
-	vrControllerEventLog.push_back(logEntry);
-	if (vrControllerEventLog.size() > 32) {
-		vrControllerEventLog.erase(vrControllerEventLog.begin());
-	}
-}
-
-void VR::UpdateControllerState(const Menu::KeyEvent& event)
-{
-	bool isPrimary = RE::BSOpenVRControllerDevice::IsPrimaryController(event.device);
-	bool isSecondary = RE::BSOpenVRControllerDevice::IsSecondaryController(event.device);
-
-	// Update thumbstick state for diagnostics display and later input processing
-	if (isPrimary) {
-		primaryControllerState.thumbsticks[static_cast<size_t>(RE::ControllerRole::Primary)].x = event.thumbstickX;
-		primaryControllerState.thumbsticks[static_cast<size_t>(RE::ControllerRole::Primary)].y = event.thumbstickY;
-	} else if (isSecondary) {
-		secondaryControllerState.thumbsticks[static_cast<size_t>(RE::ControllerRole::Secondary)].x = event.thumbstickX;
-		secondaryControllerState.thumbsticks[static_cast<size_t>(RE::ControllerRole::Secondary)].y = event.thumbstickY;
-	}
-
-	// Log the thumbstick event
-	VRControllerEventLog logEntry;
-	logEntry.device = static_cast<int>(event.device);
-	logEntry.keyCode = event.keyCode;
-	logEntry.value = static_cast<int>(event.value);
-	logEntry.pressed = event.IsPressed();
-	logEntry.heldTime = 0.0;
-	logEntry.heldSource = "thumbstick";
-	logEntry.thumbstickX = event.thumbstickX;
-	logEntry.thumbstickY = event.thumbstickY;
-	logEntry.controllerRole = isPrimary ? "Primary" : "Secondary";
-	vrControllerEventLog.push_back(logEntry);
-	if (vrControllerEventLog.size() > 32) {
-		vrControllerEventLog.erase(vrControllerEventLog.begin());
-	}
-}
-
-// Helper function to process thumbstick input for scrolling
-void VR::ProcessThumbstickScroll(RE::VRControllerState& controllerState, size_t thumbstickIndex, float deadzone, ImGuiIO& io)
-{
-	bool usingScrollStickX = (std::abs(controllerState.thumbsticks[thumbstickIndex].x) > deadzone);
-	bool usingScrollStickY = (std::abs(controllerState.thumbsticks[thumbstickIndex].y) > deadzone);
-
-	if (usingScrollStickX || usingScrollStickY) {
-		// Per-controller scroll accumulation to prevent interference between controllers
-		struct ScrollAccum
-		{
-			float x = 0.0f;
-			float y = 0.0f;
-		};
-		static std::unordered_map<size_t, ScrollAccum> scrollAccums;
-		ScrollAccum& accum = scrollAccums[thumbstickIndex];
-
-		// Accumulate scroll input with sensitivity scaling
-		accum.x += controllerState.thumbsticks[thumbstickIndex].x * 0.1f;
-		accum.y += controllerState.thumbsticks[thumbstickIndex].y * 0.1f;
-
-		// Send scroll events when accumulated enough input
-		float scrollEventX = 0.0f;
-		float scrollEventY = 0.0f;
-
-		if (std::abs(accum.x) > 0.3f) {
-			scrollEventX = accum.x > 0 ? 1.0f : -1.0f;
-			accum.x = 0.0f;
-		}
-		if (std::abs(accum.y) > 0.3f) {
-			scrollEventY = accum.y > 0 ? 1.0f : -1.0f;
-			accum.y = 0.0f;
-		}
-
-		// Send both horizontal and vertical scroll events if needed
-		// Negate X: thumbstick right (+X) should scroll content rightward,
-		// which requires negative wheel delta in ImGui convention
-		if (scrollEventX != 0.0f || scrollEventY != 0.0f) {
-			io.AddMouseWheelEvent(-scrollEventX, scrollEventY);
-		}
-	}
-}
-
-// Converts VR controller input to ImGui mouse and scroll events for the overlay UI
-// Supports both modern wand pointing and traditional thumbstick control
-void VR::ProcessControllerInputForImGui()
-{
-	if (!globals::menu->IsEnabled)
-		return;
-	bool testMode = settings.VRMenuControllerDiagnosticsTestMode;
-	float mouseDeadzone = settings.mouseDeadzone;
-	float mouseSpeed = settings.mouseSpeed;
-	ImGuiIO& io = ImGui::GetIO();
-	io.ConfigFlags &= ~ImGuiConfigFlags_NoMouseCursorChange;
-	io.WantSetMousePos = false;
-
-	// Try wand pointing first
-	bool wandHandledCursor = false;
-	if (!testMode && settings.EnableWandPointing) {
-		UpdateCursorFromWandPointing();
-		wandHandledCursor = wandState.isIntersecting;
-	}
-
-	if (!testMode) {
-		// Disable scrolling during grip drag (grip+joystick is used for depth adjustment)
-		bool isDragging = overlayDragState.dragging;
-
-		// When wand is handling cursor, BOTH thumbsticks become scroll controls
-		// When wand is NOT active, use traditional cursor/scroll assignment
-		if (wandHandledCursor && !isDragging) {
-			// Wand mode: Both thumbsticks scroll
-			ProcessThumbstickScroll(primaryControllerState, static_cast<size_t>(RE::ControllerRole::Primary), mouseDeadzone, io);
-			ProcessThumbstickScroll(secondaryControllerState, static_cast<size_t>(RE::ControllerRole::Secondary), mouseDeadzone, io);
-		} else if (!isDragging) {
-			// Traditional mode: Determine which controller handles cursor vs scrolling
-			bool useAttachedControllerForCursor = (settings.attachMode == VR::Settings::OverlayAttachMode::ControllerOnly ||
-												   settings.attachMode == VR::Settings::OverlayAttachMode::Both);
-
-			RE::VRControllerState* cursorController = nullptr;
-			RE::VRControllerState* scrollController = nullptr;
-
-			if (useAttachedControllerForCursor) {
-				// When attached to controller: attached controller = cursor, other controller = scrolling
-				if (settings.VRMenuAttachController == ControllerDevice::Primary) {
-					cursorController = &primaryControllerState;
-					scrollController = &secondaryControllerState;
-				} else {
-					cursorController = &secondaryControllerState;
-					scrollController = &primaryControllerState;
-				}
-			} else {
-				// HMD mode: primary = cursor, secondary = scroll (traditional)
-				cursorController = &primaryControllerState;
-				scrollController = &secondaryControllerState;
-			}
-
-			// Cursor movement (from determined cursor controller)
-			if (cursorController) {
-				size_t thumbstickIndex = (cursorController == &primaryControllerState) ?
-				                             static_cast<size_t>(RE::ControllerRole::Primary) :
-				                             static_cast<size_t>(RE::ControllerRole::Secondary);
-
-				float thumbstickX = cursorController->thumbsticks[thumbstickIndex].x;
-				float thumbstickY = cursorController->thumbsticks[thumbstickIndex].y;
-				bool usingCursorStick = (std::abs(thumbstickX) > mouseDeadzone || std::abs(thumbstickY) > mouseDeadzone);
-
-				if (usingCursorStick) {
-					ImVec2 mousePos = io.MousePos;
-					mousePos.x += thumbstickX * mouseSpeed;
-					mousePos.y -= thumbstickY * mouseSpeed;
-					mousePos.x = std::clamp(mousePos.x, 0.0f, io.DisplaySize.x);
-					mousePos.y = std::clamp(mousePos.y, 0.0f, io.DisplaySize.y);
-					io.MousePos = mousePos;
-					io.AddMousePosEvent(mousePos.x, mousePos.y);
-					io.MouseDrawCursor = true;
-					io.WantSetMousePos = true;
-					io.AddMouseButtonEvent(ImGuiMouseButton_Left, false);
-				}
-			}
-
-			// Scrolling (from determined scroll controller)
-			if (scrollController) {
-				size_t thumbstickIndex = (scrollController == &primaryControllerState) ?
-				                             static_cast<size_t>(RE::ControllerRole::Primary) :
-				                             static_cast<size_t>(RE::ControllerRole::Secondary);
-				ProcessThumbstickScroll(*scrollController, thumbstickIndex, mouseDeadzone, io);
-			}
-		}
-	}
-}
-
-// --- File-scope static helpers for drag logic ---
-static bool CanStartAny(vr::ETrackedControllerRole role)
-{
-	return role != vr::TrackedControllerRole_Invalid;
-}
-
-void VR::UpdateOverlayDrag()
-{
-	if (!CanPerformDrag()) {
-		return;
-	}
-
-	if (overlayDragState.dragging) {
-		UpdateActiveDrag();
-	} else {
-		TryStartNewDrag();
-	}
-}
-
-bool VR::CanPerformDrag()
-{
-	if (!settings.EnableDragToReposition)
-		return false;
-
-	// Only allow drag when CS menu is open to prevent interference with gameplay
-	if (!globals::menu->IsEnabled)
-		return false;
-
-	RE::BSOpenVR* openvr = RE::BSOpenVR::GetSingleton();
-	auto* system = openvr ? openvr->vrSystem : nullptr;
-	if (!system)
-		return false;
-
-	// Check if test mode is active - disable all dragging
-	if (settings.VRMenuControllerDiagnosticsTestMode) {
-		return false;
-	}
-
-	return true;
-}
-
-void VR::UpdateActiveDrag()
-{
-	RE::BSOpenVR* openvr = RE::BSOpenVR::GetSingleton();
-	auto* system = openvr ? openvr->vrSystem : nullptr;
-	if (!system)
-		return;
-
-	// Helper to get grip state for a controller based on actual hand position
-	auto getGripPressed = [&](bool isLeft, bool isRight) {
-		bool isLeftHanded = lastKnownLeftHandedMode;
-
-		if (isLeft) {
-			// Left hand: primary if left-handed, secondary if right-handed
-			if (isLeftHanded) {
-				return primaryControllerState[RE::BSOpenVRControllerDevice::Keys::kGrip].isPressed;
-			} else {
-				return secondaryControllerState[RE::BSOpenVRControllerDevice::Keys::kGrip].isPressed;
-			}
-		}
-		if (isRight) {
-			// Right hand: secondary if left-handed, primary if right-handed
-			if (isLeftHanded) {
-				return secondaryControllerState[RE::BSOpenVRControllerDevice::Keys::kGrip].isPressed;
-			} else {
-				return primaryControllerState[RE::BSOpenVRControllerDevice::Keys::kGrip].isPressed;
-			}
-		}
-		return false;
-	};
-
-	// Helper to reset drag state
-	auto resetDragState = [&]() {
-		overlayDragState.dragging = false;
-		overlayDragState.controllerIndex = vr::k_unTrackedDeviceIndexInvalid;
-		overlayDragState.isPrimary = false;
-		overlayDragState.isSecondary = false;
-	};
-
-	float rawMatrix[3][4];
-	if (Util::GetControllerWorldMatrix(overlayDragState.controllerIndex, rawMatrix)) {
-		vr::HmdMatrix34_t mat = Util::Float3x4ToHmdMatrix34(rawMatrix);
-		Matrix controllerMatrix = Util::HmdMatrix34ToMatrix(mat);
-
-		// Update drag based on current mode
-		switch (overlayDragState.mode) {
-		case OverlayDragState::DragMode::Controller:
-			{
-				// Get current attached controller transform to convert world deltas to local space
-				vr::TrackedDeviceIndex_t attachedControllerIndex = Util::GetControllerIndexForDevice(settings.VRMenuAttachController, lastKnownLeftHandedMode);
-
-				if (attachedControllerIndex != vr::k_unTrackedDeviceIndexInvalid) {
-					vr::TrackedDevicePose_t controllerPose;
-					if (!Util::GetDeviceToAbsoluteTrackingPoseCompatible(vr::TrackingUniverseStanding, 0, &controllerPose, 1))
-						break;
-					if (controllerPose.bPoseIsValid) {
-						Matrix attachedControllerMatrix = Util::HmdMatrix34ToMatrix(controllerPose.mDeviceToAbsoluteTracking);
-
-						// Calculate world-space delta (translation is in _41,_42,_43 after HmdMatrix34ToMatrix transpose)
-						Vector3 worldDelta(
-							controllerMatrix._41 - overlayDragState.initialControllerMatrix._41,
-							controllerMatrix._42 - overlayDragState.initialControllerMatrix._42,
-							controllerMatrix._43 - overlayDragState.initialControllerMatrix._43);
-
-						// Transform world delta to attached controller local space
-						Matrix worldToLocal = attachedControllerMatrix.Invert();
-						Vector3 localDelta = Vector3::TransformNormal(worldDelta, worldToLocal);
-
-						// Apply local delta to offsets
-						settings.VRMenuControllerOffsetX = overlayDragState.initialControllerOffset.x + localDelta.x;
-						settings.VRMenuControllerOffsetY = overlayDragState.initialControllerOffset.y + localDelta.y;
-						settings.VRMenuControllerOffsetZ = overlayDragState.initialControllerOffset.z + localDelta.z;
-					}
-				}
-				break;
-			}
-		case OverlayDragState::DragMode::FixedWorld:
-			{
-				// Translation-only drag: move overlay position without inheriting wrist rotation
-				Vector3 worldDelta(
-					controllerMatrix._41 - overlayDragState.initialControllerMatrix._41,
-					controllerMatrix._42 - overlayDragState.initialControllerMatrix._42,
-					controllerMatrix._43 - overlayDragState.initialControllerMatrix._43);
-				Matrix translated = overlayDragState.initialOverlayMatrix;
-				translated._41 += worldDelta.x;
-				translated._42 += worldDelta.y;
-				translated._43 += worldDelta.z;
-				fixedWorldOverlayPosition.m = translated;
-				break;
-			}
-		case OverlayDragState::DragMode::HMD:
-			{
-				// Get current HMD transform to convert world deltas to local space
-				vr::TrackedDevicePose_t hmdPose;
-				if (!Util::GetDeviceToAbsoluteTrackingPoseCompatible(vr::TrackingUniverseStanding, 0, &hmdPose, 1))
-					break;
-				if (hmdPose.bPoseIsValid) {
-					Matrix hmdMatrix = Util::HmdMatrix34ToMatrix(hmdPose.mDeviceToAbsoluteTracking);
-
-					// Calculate world-space delta
-					// Note: HmdMatrix34ToMatrix converts to Row-Major Row-Vector format, so translation is in the bottom row (_41, _42, _43)
-					Vector3 worldDelta(
-						controllerMatrix._41 - overlayDragState.initialControllerMatrix._41,
-						controllerMatrix._42 - overlayDragState.initialControllerMatrix._42,
-						controllerMatrix._43 - overlayDragState.initialControllerMatrix._43);
-
-					// Transform world delta to HMD local space
-					// Use Invert to convert World Delta -> Local Delta
-					// Use TransformNormal to treat as vector (ignore translation)
-					Matrix worldToLocal = hmdMatrix.Invert();
-					Vector3 localDelta = Vector3::TransformNormal(worldDelta, worldToLocal);
-
-					// Debug: Log delta values
-					static auto lastDeltaLog = std::chrono::steady_clock::now();
-					auto nowDelta = std::chrono::steady_clock::now();
-					if (std::chrono::duration_cast<std::chrono::milliseconds>(nowDelta - lastDeltaLog).count() > 500) {
-						logger::debug("VR Drag Delta - Local: ({:.3f}, {:.3f}, {:.3f})", localDelta.x, localDelta.y, localDelta.z);
-						lastDeltaLog = nowDelta;
-					}
-
-					// Apply X/Y delta to position offsets
-					settings.VRMenuOffsetX = overlayDragState.initialHMDOffset.x + localDelta.x;
-					settings.VRMenuOffsetY = overlayDragState.initialHMDOffset.y + localDelta.y;
-
-					// Z delta controls actual depth (VRMenuOffsetZ)
-					// Moving controller forward (-Z) should push menu away (more negative Z)
-					// Use + localDelta.z because forward movement is negative Z, which makes offset more negative (further away)
-					settings.VRMenuOffsetZ = overlayDragState.initialHMDOffset.z + localDelta.z;
-
-					// Scale remains constant during drag (can be adjusted in settings if needed)
-					settings.VRMenuScale = overlayDragState.initialHMDScale;
-
-					static std::chrono::steady_clock::time_point lastLog = std::chrono::steady_clock::now();
-					auto now = std::chrono::steady_clock::now();
-					if (std::chrono::duration_cast<std::chrono::milliseconds>(now - lastLog).count() > 500) {
-						logger::debug("VR Dragging (3D Mode): Offset ({:.2f}, {:.2f}, {:.2f}), Scale {:.2f}",
-							settings.VRMenuOffsetX, settings.VRMenuOffsetY, settings.VRMenuOffsetZ, settings.VRMenuScale);
-						lastLog = now;
-					}
-				}
-				break;
-			}
-		}
-	}
-
-	// Joystick depth control during grip: thumbstick Y on gripping controller adjusts depth
-	if (overlayDragState.dragging) {
-		// Determine which controller state to read based on which hand is gripping
-		// isPrimary/isSecondary refer to left/right hand, map to primary/secondary controller state
-		RE::VRControllerState* gripController = nullptr;
-		size_t thumbIdx = 0;
-		if (overlayDragState.isPrimary) {
-			// Left hand is gripping
-			if (lastKnownLeftHandedMode) {
-				gripController = &primaryControllerState;
-				thumbIdx = static_cast<size_t>(RE::ControllerRole::Primary);
-			} else {
-				gripController = &secondaryControllerState;
-				thumbIdx = static_cast<size_t>(RE::ControllerRole::Secondary);
-			}
-		} else if (overlayDragState.isSecondary) {
-			// Right hand is gripping
-			if (lastKnownLeftHandedMode) {
-				gripController = &secondaryControllerState;
-				thumbIdx = static_cast<size_t>(RE::ControllerRole::Secondary);
-			} else {
-				gripController = &primaryControllerState;
-				thumbIdx = static_cast<size_t>(RE::ControllerRole::Primary);
-			}
-		}
-
-		if (gripController) {
-			float thumbY = gripController->thumbsticks[thumbIdx].y;
-			const float deadzone = settings.mouseDeadzone;
-			const float depthSpeed = 0.02f;
-			if (std::abs(thumbY) > deadzone) {
-				// Negative thumbY (stick forward/up) = move away (+Z depth)
-				// Positive thumbY (stick back/down) = bring closer (-Z depth)
-				float depthDelta = -thumbY * depthSpeed;
-				if (overlayDragState.mode == OverlayDragState::DragMode::HMD) {
-					// Accumulate into the initial offset so the drag code doesn't overwrite it
-					overlayDragState.initialHMDOffset.z += depthDelta;
-				} else if (overlayDragState.mode == OverlayDragState::DragMode::Controller) {
-					overlayDragState.initialControllerOffset.z += depthDelta;
-				}
-			}
-		}
-	}
-
-	bool gripPressed = getGripPressed(overlayDragState.isPrimary, overlayDragState.isSecondary);
-	if (!gripPressed) {
-		resetDragState();
-	}
-}
-
-void VR::TryStartNewDrag()
-{
-	RE::BSOpenVR* openvr = RE::BSOpenVR::GetSingleton();
-	auto* system = openvr ? openvr->vrSystem : nullptr;
-	if (!system)
-		return;
-
-	// Helper to get grip state for a controller based on actual hand position
-	auto getGripPressed = [&](bool isLeft, bool isRight) {
-		bool isLeftHanded = lastKnownLeftHandedMode;
-
-		if (isLeft) {
-			// Left hand: primary if left-handed, secondary if right-handed
-			if (isLeftHanded) {
-				return primaryControllerState[RE::BSOpenVRControllerDevice::Keys::kGrip].isPressed;
-			} else {
-				return secondaryControllerState[RE::BSOpenVRControllerDevice::Keys::kGrip].isPressed;
-			}
-		}
-		if (isRight) {
-			// Right hand: secondary if left-handed, primary if right-handed
-			if (isLeftHanded) {
-				return secondaryControllerState[RE::BSOpenVRControllerDevice::Keys::kGrip].isPressed;
-			} else {
-				return primaryControllerState[RE::BSOpenVRControllerDevice::Keys::kGrip].isPressed;
-			}
-		}
-		return false;
-	};
-
-	// --- Strict mutually exclusive drag mode selection ---
-	struct DragMode
-	{
-		OverlayDragState::DragMode mode;
-		bool isActive;
-		std::function<bool(vr::ETrackedControllerRole)> canStart;
-		std::function<void()> onInit;
-	};
-
-	std::vector<DragMode> dragModes;
-
-	// Controller mode - only for opposite hand (highest priority)
-	if (settings.attachMode == AttachMode::ControllerOnly || settings.attachMode == AttachMode::Both) {
-		// Controller drag - only opposite hand can drag the controller overlay
-		dragModes.push_back({ OverlayDragState::DragMode::Controller,
-			true,
-			[&](vr::ETrackedControllerRole role) {
-				// Get the attached controller index
-				vr::TrackedDeviceIndex_t attachedControllerIndex = Util::GetControllerIndexForDevice(settings.VRMenuAttachController, lastKnownLeftHandedMode);
-				if (attachedControllerIndex == vr::k_unTrackedDeviceIndexInvalid) {
-					return false;  // No attached controller found
-				}
-
-				// Get the opposite controller index
-				ControllerDevice oppositeDevice = (settings.VRMenuAttachController == ControllerDevice::Primary) ?
-			                                          ControllerDevice::Secondary :
-			                                          ControllerDevice::Primary;
-				vr::TrackedDeviceIndex_t oppositeControllerIndex = Util::GetControllerIndexForDevice(oppositeDevice, lastKnownLeftHandedMode);
-				if (oppositeControllerIndex == vr::k_unTrackedDeviceIndexInvalid) {
-					return false;  // No opposite controller found
-				}
-
-				// Check if the current controller (the one doing the dragging) is the opposite controller
-				for (vr::TrackedDeviceIndex_t i = 0; i < vr::k_unMaxTrackedDeviceCount; ++i) {
-					if (system->GetTrackedDeviceClass(i) == vr::TrackedDeviceClass_Controller) {
-						vr::ETrackedControllerRole deviceRole = system->GetControllerRoleForTrackedDeviceIndex(i);
-						if (deviceRole == role && i == oppositeControllerIndex) {
-							return true;  // This is the opposite controller, can drag
-						}
-					}
-				}
-				return false;  // Not the opposite controller
-			},
-			[&]() {
-				overlayDragState.initialControllerOffset.x = settings.VRMenuControllerOffsetX;
-				overlayDragState.initialControllerOffset.y = settings.VRMenuControllerOffsetY;
-				overlayDragState.initialControllerOffset.z = settings.VRMenuControllerOffsetZ;
-				overlayDragState.initialControllerMatrix = overlayDragState.startControllerMatrix;
-			} });
-	}
-
-	// Fixed world mode - only for attached controller when in "Both" mode
-	if (settings.VRMenuPositioningMethod == 1) {
-		// In "Both" mode, only the attached controller can adjust fixed world position
-		// In HMD-only mode, any controller can adjust fixed world position
-		std::function<bool(vr::ETrackedControllerRole)> fixedWorldCanStart;
-		if (settings.attachMode == AttachMode::Both) {
-			// In "Both" mode, only the attached controller can adjust fixed world position
-			fixedWorldCanStart = [&](vr::ETrackedControllerRole role) {
-				// Find the actual attached controller using helper function
-				vr::TrackedDeviceIndex_t attachedControllerIndex = Util::GetControllerIndexForDevice(settings.VRMenuAttachController, lastKnownLeftHandedMode);
-
-				if (attachedControllerIndex != vr::k_unTrackedDeviceIndexInvalid) {
-					vr::ETrackedControllerRole actualAttachedRole = system->GetControllerRoleForTrackedDeviceIndex(attachedControllerIndex);
-					// Only allow the attached controller to drag fixed world
-					return role == actualAttachedRole;
-				}
-				return false;
-			};
-		} else {
-			// In HMD-only mode, any controller can adjust fixed world position
-			fixedWorldCanStart = CanStartAny;
-		}
-
-		dragModes.push_back({ OverlayDragState::DragMode::FixedWorld,
-			true,
-			fixedWorldCanStart,
-			[&]() {
-				overlayDragState.initialControllerMatrix = overlayDragState.startControllerMatrix;
-				overlayDragState.initialOverlayMatrix = fixedWorldOverlayPosition.m;
-			} });
-	}
-
-	// HMD mode - for attached controller when both modes active, or any controller otherwise
-	if (settings.attachMode == AttachMode::HMDOnly || settings.attachMode == AttachMode::Both) {
-		// In "Both" mode, only the attached controller can adjust HMD position
-		// In HMD-only mode, any controller can adjust HMD position
-		std::function<bool(vr::ETrackedControllerRole)> hmdCanStart;
-		if (settings.attachMode == AttachMode::Both) {
-			// In "Both" mode, only the attached controller can adjust HMD position
-			hmdCanStart = [&](vr::ETrackedControllerRole role) {
-				// Find the actual attached controller using helper function
-				vr::TrackedDeviceIndex_t attachedControllerIndex = Util::GetControllerIndexForDevice(settings.VRMenuAttachController, lastKnownLeftHandedMode);
-
-				if (attachedControllerIndex != vr::k_unTrackedDeviceIndexInvalid) {
-					vr::ETrackedControllerRole actualAttachedRole = system->GetControllerRoleForTrackedDeviceIndex(attachedControllerIndex);
-					// Only allow the attached controller to drag HMD
-					return role == actualAttachedRole;
-				}
-				return false;
-			};
-		} else {
-			// In HMD-only mode, any controller can adjust HMD
-			hmdCanStart = CanStartAny;
-		}
-
-		dragModes.push_back({ OverlayDragState::DragMode::HMD,
-			true,
-			hmdCanStart,
-			[&]() {
-				overlayDragState.initialHMDOffset.x = settings.VRMenuOffsetX;
-				overlayDragState.initialHMDOffset.y = settings.VRMenuOffsetY;
-				overlayDragState.initialHMDOffset.z = settings.VRMenuOffsetZ;
-				overlayDragState.initialHMDScale = settings.VRMenuScale;
-				overlayDragState.initialControllerMatrix = overlayDragState.startControllerMatrix;
-			} });
-	}
-
-	// Try to start a new drag - use first available mode
-	for (const auto& mode : dragModes) {
-		if (!mode.isActive)
-			continue;
-		for (vr::TrackedDeviceIndex_t i = 0; i < vr::k_unMaxTrackedDeviceCount; ++i) {
-			if (system->GetTrackedDeviceClass(i) != vr::TrackedDeviceClass_Controller)
-				continue;
-			vr::ETrackedControllerRole role = system->GetControllerRoleForTrackedDeviceIndex(i);
-			bool isLeft = (role == vr::ETrackedControllerRole::TrackedControllerRole_LeftHand);
-			bool isRight = (role == vr::ETrackedControllerRole::TrackedControllerRole_RightHand);
-			if (!mode.canStart(role))
-				continue;
-			bool gripPressed = getGripPressed(isLeft, isRight);
-			if (!gripPressed)
-				continue;
-			float rawMatrix[3][4];
-			if (!Util::GetControllerWorldMatrix(i, rawMatrix))
-				continue;
-			vr::HmdMatrix34_t mat = Util::Float3x4ToHmdMatrix34(rawMatrix);
-			Matrix controllerMatrix = Util::HmdMatrix34ToMatrix(mat);
-			overlayDragState.dragging = true;
-			overlayDragState.mode = mode.mode;
-			overlayDragState.controllerIndex = i;
-			overlayDragState.isPrimary = isLeft;
-			overlayDragState.isSecondary = isRight;
-			overlayDragState.startControllerMatrix = controllerMatrix;
-			mode.onInit();
-
-			// Send haptic pulse to the controller that started the drag (only if overlay is visible)
-			if (system && globals::menu->IsEnabled) {
-				// Find the controller device index for the hand that started the drag
-				for (vr::TrackedDeviceIndex_t deviceIdx = 0; deviceIdx < vr::k_unMaxTrackedDeviceCount; ++deviceIdx) {
-					if (system->GetTrackedDeviceClass(deviceIdx) == vr::TrackedDeviceClass_Controller) {
-						vr::ETrackedControllerRole deviceRole = system->GetControllerRoleForTrackedDeviceIndex(deviceIdx);
-						bool isRightController = (deviceRole == vr::ETrackedControllerRole::TrackedControllerRole_RightHand);
-						if (isRightController == isRight) {
-							// Use BSOpenVR's haptic pulse method instead of direct OpenVR call
-							openvr->TriggerHapticPulse(isRightController, 25.0f);  // 100ms pulse
-							break;
-						}
-					}
-				}
-			}
-
-			return;
-		}
-	}
-}
-
-void VR::SetFixedOverlayToCurrentHMD()
-{
-	vr::HmdMatrix34_t transform = Util::ComputeOverlayTransformFromHMD(
-		settings.VRMenuOffsetX,
-		settings.VRMenuOffsetY,
-		settings.VRMenuOffsetZ);
-	fixedWorldOverlayPosition.m = Util::HmdMatrix34ToMatrix(transform);
-}
-
-void VR::UpdateFixedWorldPositioning()
-{
-	if (settings.VRMenuPositioningMethod != 1)
-		return;
-
-	// Initialize fixed world position on first use (identity matrix means uninitialized)
-	bool isUninitialized = (fixedWorldOverlayPosition.m._41 == 0.0f &&
-							fixedWorldOverlayPosition.m._42 == 0.0f &&
-							fixedWorldOverlayPosition.m._43 == 0.0f &&
-							fixedWorldOverlayPosition.m._11 == 1.0f);
-
-	if (isUninitialized) {
-		SetFixedOverlayToCurrentHMD();
-		auto player = RE::PlayerCharacter::GetSingleton();
-		if (player) {
-			savedPlayerWorldPos = player->GetPosition();
-		}
-		return;
-	}
-
-	// Auto-reset: only if player has moved beyond threshold from where overlay was last placed
-	if (settings.VRMenuAutoResetDistance > 0.0f) {
-		auto player = RE::PlayerCharacter::GetSingleton();
-		if (player) {
-			RE::NiPoint3 playerPos = player->GetPosition();
-			float sqDist = playerPos.GetSquaredDistance(savedPlayerWorldPos);
-			float thresholdSq = settings.VRMenuAutoResetDistance * settings.VRMenuAutoResetDistance;
-			if (sqDist > thresholdSq) {
-				SetFixedOverlayToCurrentHMD();
-				savedPlayerWorldPos = playerPos;
-			}
 		}
 	}
 }
@@ -2302,159 +246,4 @@ void VR::DetectOpenVRInfo()
 bool VR::IsOpenVRCompatible() const
 {
 	return openVRInfo.isCompatible;
-}
-
-//=============================================================================
-// WAND POINTING IMPLEMENTATION
-//=============================================================================
-
-bool VR::ComputeWandIntersectionForOverlayType(OverlayType type, vr::TrackedDeviceIndex_t controllerIndex, ImVec2& outUV)
-{
-	// 1. Get Controller Ray
-	float controllerM[3][4];
-	if (!Util::GetControllerWorldMatrix(controllerIndex, controllerM)) {
-		return false;
-	}
-	Matrix controllerWorld = Util::HmdMatrix34ToMatrix(Util::Float3x4ToHmdMatrix34(controllerM));
-	Vector3 rayOrigin = controllerWorld.Translation();
-	Vector3 rayDir = controllerWorld.Forward();
-
-	// 2. Get Overlay World Matrix
-	Matrix overlayWorld;
-	if (type == OverlayType::HMD) {
-		if (settings.VRMenuPositioningMethod == 1) {  // Fixed
-			overlayWorld = fixedWorldOverlayPosition.m;
-		} else {  // HMD Relative
-			vr::TrackedDevicePose_t hmdPose;
-			if (!Util::GetDeviceToAbsoluteTrackingPoseCompatible(vr::TrackingUniverseStanding, 0, &hmdPose, 1))
-				return false;
-			if (!hmdPose.bPoseIsValid)
-				return false;
-			Matrix hmdWorld = Util::HmdMatrix34ToMatrix(hmdPose.mDeviceToAbsoluteTracking);
-			Matrix offset = Matrix::CreateTranslation(settings.VRMenuOffsetX, settings.VRMenuOffsetY, settings.VRMenuOffsetZ);
-			overlayWorld = offset * hmdWorld;
-		}
-	} else {  // Controller Relative
-		vr::TrackedDeviceIndex_t attachIndex = Util::GetControllerIndexForDevice(settings.VRMenuAttachController, lastKnownLeftHandedMode);
-		if (attachIndex == vr::k_unTrackedDeviceIndexInvalid)
-			return false;
-
-		float attachM[3][4];
-		if (!Util::GetControllerWorldMatrix(attachIndex, attachM))
-			return false;
-		Matrix attachWorld = Util::HmdMatrix34ToMatrix(Util::Float3x4ToHmdMatrix34(attachM));
-
-		Matrix offset = Matrix::CreateTranslation(settings.VRMenuControllerOffsetX, settings.VRMenuControllerOffsetY, settings.VRMenuControllerOffsetZ);
-		overlayWorld = offset * attachWorld;
-	}
-
-	// Apply Scale
-	overlayWorld = Matrix::CreateScale(settings.VRMenuScale) * overlayWorld;
-
-	// 3. Transform Ray to Local Space
-	Matrix worldToOverlay = overlayWorld.Invert();
-	Vector3 localOrigin = Vector3::Transform(rayOrigin, worldToOverlay);
-	Vector3 localDir = Vector3::TransformNormal(rayDir, worldToOverlay);
-
-	// 4. Intersect with Plane Z=0
-	if (abs(localDir.z) < 1e-6f)
-		return false;  // Parallel
-
-	float t = -localOrigin.z / localDir.z;
-	if (t < 0.0f)
-		return false;  // Behind ray
-
-	// Intersection Point
-	Vector3 hit = localOrigin + t * localDir;
-
-	// 5. Check Bounds (-0.5 to 0.5)
-	if (hit.x < -0.5f || hit.x > 0.5f || hit.y < -0.5f || hit.y > 0.5f)
-		return false;
-
-	// 6. Compute UV
-	// Quad spans [-0.5, 0.5]. UV (0,0) is top-left of texture (standard D3D).
-	// +Y on quad = top = UV V=0, so V = 0.5 - hit.y
-	outUV.x = hit.x + 0.5f;
-	outUV.y = 0.5f - hit.y;
-
-	return true;
-}
-
-bool VR::ComputeWandIntersection(vr::TrackedDeviceIndex_t controllerIndex, ImVec2& outUV)
-{
-	// Try HMD-attached overlay first, then controller-attached based on settings
-	bool intersected = false;
-	if (settings.attachMode == AttachMode::HMDOnly || settings.attachMode == AttachMode::Both) {
-		if (ComputeWandIntersectionForOverlayType(OverlayType::HMD, controllerIndex, outUV)) {
-			intersected = true;
-		}
-	}
-	if (!intersected && (settings.attachMode == AttachMode::ControllerOnly || settings.attachMode == AttachMode::Both)) {
-		if (ComputeWandIntersectionForOverlayType(OverlayType::Controller, controllerIndex, outUV)) {
-			intersected = true;
-		}
-	}
-
-	if (intersected) {
-		wandState.isIntersecting = true;
-		wandState.uvCoordinates = outUV;
-		wandState.controllerIndex = controllerIndex;
-	} else {
-		wandState.isIntersecting = false;
-	}
-
-	return intersected;
-}
-
-void VR::UpdateCursorFromWandPointing()
-{
-	// Only update cursor when menu is active (ImGui must be initialized)
-	if (!settings.EnableWandPointing || !globals::menu->IsEnabled)
-		return;
-
-	ImGuiIO& io = ImGui::GetIO();
-
-	// Determine which controller should be used for pointing
-	// Use non-attached controller (free hand points at menu on other hand)
-	vr::TrackedDeviceIndex_t pointingController = vr::k_unTrackedDeviceIndexInvalid;
-
-	if (settings.attachMode == AttachMode::ControllerOnly || settings.attachMode == AttachMode::Both) {
-		// If menu is attached to a controller, use the OTHER controller for pointing
-		ControllerDevice oppositeController = (settings.VRMenuAttachController == ControllerDevice::Primary) ?
-		                                          ControllerDevice::Secondary :
-		                                          ControllerDevice::Primary;
-		pointingController = Util::GetControllerIndexForDevice(oppositeController, lastKnownLeftHandedMode);
-	} else {
-		// HMD-only mode: use primary (dominant) hand for pointing
-		pointingController = Util::GetControllerIndexForDevice(ControllerDevice::Primary, lastKnownLeftHandedMode);
-	}
-
-	if (pointingController == vr::k_unTrackedDeviceIndexInvalid) {
-		wandState.isIntersecting = false;
-		return;
-	}
-
-	// Compute wand intersection using ray-quad intersection
-	ImVec2 uv;
-	bool intersected = ComputeWandIntersection(pointingController, uv);
-
-	// Update cursor position from intersection
-	if (intersected) {
-		// Convert UV (0-1) to screen pixel coordinates
-		// UV already in screen convention: (0,0) = top-left, (1,1) = bottom-right
-		float screenX = uv.x * io.DisplaySize.x;
-		float screenY = uv.y * io.DisplaySize.y;
-
-		// Clamp to screen bounds (allow full display size, consistent with thumbstick cursor)
-		screenX = std::clamp(screenX, 0.0f, io.DisplaySize.x);
-		screenY = std::clamp(screenY, 0.0f, io.DisplaySize.y);
-
-		io.MousePos = ImVec2(screenX, screenY);
-		io.AddMousePosEvent(screenX, screenY);
-		io.MouseDrawCursor = true;
-		io.WantSetMousePos = true;
-	} else {
-		// Ensure wand state is cleared if no intersection
-		wandState.isIntersecting = false;
-	}
 }
